move webpack config files, fix paths, remove jquery, fix doc build
https://coveord.atlassian.net/browse/JSUI-1146
