+++ gulpTasks/doc.js
@@ -30,7 +30,7 @@ gulp.task('copyBinToDoc', function () {
 })
 
 gulp.task('buildPlayground', shell.task([
-  'node node_modules/webpack/bin/webpack.js --config webpack.playground.config.js'
+  'node node_modules/webpack/bin/webpack.js --config webpackConfigFiles/webpack.playground.config.js'
 ]));
 
 function copyFile(source, target, cb) {
