--- typings/browser/ambient/jquery/index.d.ts
@@ -1,3223 +1,0 @@
-// Generated by typings
-// Source: https://raw.githubusercontent.com/DefinitelyTyped/DefinitelyTyped/4cdfbe96b666eec5e1defbb519a62abf04e96764/jquery/jquery.d.ts
-// Type definitions for jQuery 1.10.x / 2.0.x
-// Project: http://jquery.com/
-// Definitions by: Boris Yankov <https://github.com/borisyankov/>, Christian Hoffmeister <https://github.com/choffmeister>, Steve Fenton <https://github.com/Steve-Fenton>, Diullei Gomes <https://github.com/Diullei>, Tass Iliopoulos <https://github.com/tasoili>, Jason Swearingen <https://github.com/jasons-novaleaf>, Sean Hill <https://github.com/seanski>, Guus Goossens <https://github.com/Guuz>, Kelly Summerlin <https://github.com/ksummerlin>, Basarat Ali Syed <https://github.com/basarat>, Nicholas Wolverson <https://github.com/nwolverson>, Derek Cicerone <https://github.com/derekcicerone>, Andrew Gaspar <https://github.com/AndrewGaspar>, James Harrison Fisher <https://github.com/jameshfisher>, Seikichi Kondo <https://github.com/seikichi>, Benjamin Jackman <https://github.com/benjaminjackman>, Poul Sorensen <https://github.com/s093294>, Josh Strobl <https://github.com/JoshStrobl>, John Reilly <https://github.com/johnnyreilly/>, Dick van den Brink <https://github.com/DickvdBrink>
-// Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
-
-/* *****************************************************************************
-Copyright (c) Microsoft Corporation. All rights reserved.
-Licensed under the Apache License, Version 2.0 (the "License"); you may not use
-this file except in compliance with the License. You may obtain a copy of the
-License at http://www.apache.org/licenses/LICENSE-2.0
-
-THIS CODE IS PROVIDED *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION ANY IMPLIED
-WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A PARTICULAR PURPOSE,
-MERCHANTABLITY OR NON-INFRINGEMENT.
-
-See the Apache Version 2.0 License for specific language governing permissions
-and limitations under the License.
-***************************************************************************** */
-
-
-/**
- * Interface for the AJAX setting that will configure the AJAX request
- */
-interface JQueryAjaxSettings {
-    /**
-     * The content type sent in the request header that tells the server what kind of response it will accept in return. If the accepts setting needs modification, it is recommended to do so once in the $.ajaxSetup() method.
-     */
-    accepts?: any;
-    /**
-     * By default, all requests are sent asynchronously (i.e. this is set to true by default). If you need synchronous requests, set this option to false. Cross-domain requests and dataType: "jsonp" requests do not support synchronous operation. Note that synchronous requests may temporarily lock the browser, disabling any actions while the request is active. As of jQuery 1.8, the use of async: false with jqXHR ($.Deferred) is deprecated; you must use the success/error/complete callback options instead of the corresponding methods of the jqXHR object such as jqXHR.done() or the deprecated jqXHR.success().
-     */
-    async?: boolean;
-    /**
-     * A pre-request callback function that can be used to modify the jqXHR (in jQuery 1.4.x, XMLHTTPRequest) object before it is sent. Use this to set custom headers, etc. The jqXHR and settings objects are passed as arguments. This is an Ajax Event. Returning false in the beforeSend function will cancel the request. As of jQuery 1.5, the beforeSend option will be called regardless of the type of request.
-     */
-    beforeSend? (jqXHR: JQueryXHR, settings: JQueryAjaxSettings): any;
-    /**
-     * If set to false, it will force requested pages not to be cached by the browser. Note: Setting cache to false will only work correctly with HEAD and GET requests. It works by appending "_={timestamp}" to the GET parameters. The parameter is not needed for other types of requests, except in IE8 when a POST is made to a URL that has already been requested by a GET.
-     */
-    cache?: boolean;
-    /**
-     * A function to be called when the request finishes (after success and error callbacks are executed). The function gets passed two arguments: The jqXHR (in jQuery 1.4.x, XMLHTTPRequest) object and a string categorizing the status of the request ("success", "notmodified", "error", "timeout", "abort", or "parsererror"). As of jQuery 1.5, the complete setting can accept an array of functions. Each function will be called in turn. This is an Ajax Event.
-     */
-    complete? (jqXHR: JQueryXHR, textStatus: string): any;
-    /**
-     * An object of string/regular-expression pairs that determine how jQuery will parse the response, given its content type. (version added: 1.5)
-     */
-    contents?: { [key: string]: any; };
-    //According to jQuery.ajax source code, ajax's option actually allows contentType to set to "false"
-    // https://github.com/DefinitelyTyped/DefinitelyTyped/issues/742
-    /**
-     * When sending data to the server, use this content type. Default is "application/x-www-form-urlencoded; charset=UTF-8", which is fine for most cases. If you explicitly pass in a content-type to $.ajax(), then it is always sent to the server (even if no data is sent). The W3C XMLHttpRequest specification dictates that the charset is always UTF-8; specifying another charset will not force the browser to change the encoding.
-     */
-    contentType?: any;
-    /**
-     * This object will be made the context of all Ajax-related callbacks. By default, the context is an object that represents the ajax settings used in the call ($.ajaxSettings merged with the settings passed to $.ajax).
-     */
-    context?: any;
-    /**
-     * An object containing dataType-to-dataType converters. Each converter's value is a function that returns the transformed value of the response. (version added: 1.5)
-     */
-    converters?: { [key: string]: any; };
-    /**
-     * If you wish to force a crossDomain request (such as JSONP) on the same domain, set the value of crossDomain to true. This allows, for example, server-side redirection to another domain. (version added: 1.5)
-     */
-    crossDomain?: boolean;
-    /**
-     * Data to be sent to the server. It is converted to a query string, if not already a string. It's appended to the url for GET-requests. See processData option to prevent this automatic processing. Object must be Key/Value pairs. If value is an Array, jQuery serializes multiple values with same key based on the value of the traditional setting (described below).
-     */
-    data?: any;
-    /**
-     * A function to be used to handle the raw response data of XMLHttpRequest.This is a pre-filtering function to sanitize the response. You should return the sanitized data. The function accepts two arguments: The raw data returned from the server and the 'dataType' parameter.
-     */
-    dataFilter? (data: any, ty: any): any;
-    /**
-     * The type of data that you're expecting back from the server. If none is specified, jQuery will try to infer it based on the MIME type of the response (an XML MIME type will yield XML, in 1.4 JSON will yield a JavaScript object, in 1.4 script will execute the script, and anything else will be returned as a string). 
-     */
-    dataType?: string;
-    /**
-     * A function to be called if the request fails. The function receives three arguments: The jqXHR (in jQuery 1.4.x, XMLHttpRequest) object, a string describing the type of error that occurred and an optional exception object, if one occurred. Possible values for the second argument (besides null) are "timeout", "error", "abort", and "parsererror". When an HTTP error occurs, errorThrown receives the textual portion of the HTTP status, such as "Not Found" or "Internal Server Error." As of jQuery 1.5, the error setting can accept an array of functions. Each function will be called in turn. Note: This handler is not called for cross-domain script and cross-domain JSONP requests. This is an Ajax Event.
-     */
-    error? (jqXHR: JQueryXHR, textStatus: string, errorThrown: string): any;
-    /**
-     * Whether to trigger global Ajax event handlers for this request. The default is true. Set to false to prevent the global handlers like ajaxStart or ajaxStop from being triggered. This can be used to control various Ajax Events.
-     */
-    global?: boolean;
-    /**
-     * An object of additional header key/value pairs to send along with requests using the XMLHttpRequest transport. The header X-Requested-With: XMLHttpRequest is always added, but its default XMLHttpRequest value can be changed here. Values in the headers setting can also be overwritten from within the beforeSend function. (version added: 1.5)
-     */
-    headers?: { [key: string]: any; };
-    /**
-     * Allow the request to be successful only if the response has changed since the last request. This is done by checking the Last-Modified header. Default value is false, ignoring the header. In jQuery 1.4 this technique also checks the 'etag' specified by the server to catch unmodified data.
-     */
-    ifModified?: boolean;
-    /**
-     * Allow the current environment to be recognized as "local," (e.g. the filesystem), even if jQuery does not recognize it as such by default. The following protocols are currently recognized as local: file, *-extension, and widget. If the isLocal setting needs modification, it is recommended to do so once in the $.ajaxSetup() method. (version added: 1.5.1)
-     */
-    isLocal?: boolean;
-    /**
-     * Override the callback function name in a jsonp request. This value will be used instead of 'callback' in the 'callback=?' part of the query string in the url. So {jsonp:'onJSONPLoad'} would result in 'onJSONPLoad=?' passed to the server. As of jQuery 1.5, setting the jsonp option to false prevents jQuery from adding the "?callback" string to the URL or attempting to use "=?" for transformation. In this case, you should also explicitly set the jsonpCallback setting. For example, { jsonp: false, jsonpCallback: "callbackName" }
-     */
-    jsonp?: any;
-    /**
-     * Specify the callback function name for a JSONP request. This value will be used instead of the random name automatically generated by jQuery. It is preferable to let jQuery generate a unique name as it'll make it easier to manage the requests and provide callbacks and error handling. You may want to specify the callback when you want to enable better browser caching of GET requests. As of jQuery 1.5, you can also use a function for this setting, in which case the value of jsonpCallback is set to the return value of that function.
-     */
-    jsonpCallback?: any;
-    /**
-     * The HTTP method to use for the request (e.g. "POST", "GET", "PUT"). (version added: 1.9.0)
-     */
-    method?: string;
-    /**
-     * A mime type to override the XHR mime type. (version added: 1.5.1)
-     */
-    mimeType?: string;
-    /**
-     * A password to be used with XMLHttpRequest in response to an HTTP access authentication request.
-     */
-    password?: string;
-    /**
-     * By default, data passed in to the data option as an object (technically, anything other than a string) will be processed and transformed into a query string, fitting to the default content-type "application/x-www-form-urlencoded". If you want to send a DOMDocument, or other non-processed data, set this option to false.
-     */
-    processData?: boolean;
-    /**
-     * Only applies when the "script" transport is used (e.g., cross-domain requests with "jsonp" or "script" dataType and "GET" type). Sets the charset attribute on the script tag used in the request. Used when the character set on the local page is not the same as the one on the remote script.
-     */
-    scriptCharset?: string;
-    /**
-     * An object of numeric HTTP codes and functions to be called when the response has the corresponding code. f the request is successful, the status code functions take the same parameters as the success callback; if it results in an error (including 3xx redirect), they take the same parameters as the error callback. (version added: 1.5)
-     */
-    statusCode?: { [key: string]: any; };
-    /**
-     * A function to be called if the request succeeds. The function gets passed three arguments: The data returned from the server, formatted according to the dataType parameter; a string describing the status; and the jqXHR (in jQuery 1.4.x, XMLHttpRequest) object. As of jQuery 1.5, the success setting can accept an array of functions. Each function will be called in turn. This is an Ajax Event.
-     */
-    success? (data: any, textStatus: string, jqXHR: JQueryXHR): any;
-    /**
-     * Set a timeout (in milliseconds) for the request. This will override any global timeout set with $.ajaxSetup(). The timeout period starts at the point the $.ajax call is made; if several other requests are in progress and the browser has no connections available, it is possible for a request to time out before it can be sent. In jQuery 1.4.x and below, the XMLHttpRequest object will be in an invalid state if the request times out; accessing any object members may throw an exception. In Firefox 3.0+ only, script and JSONP requests cannot be cancelled by a timeout; the script will run even if it arrives after the timeout period.
-     */
-    timeout?: number;
-    /**
-     * Set this to true if you wish to use the traditional style of param serialization.
-     */
-    traditional?: boolean;
-    /**
-     * The type of request to make ("POST" or "GET"), default is "GET". Note: Other HTTP request methods, such as PUT and DELETE, can also be used here, but they are not supported by all browsers.
-     */
-    type?: string;
-    /**
-     * A string containing the URL to which the request is sent.
-     */
-    url?: string;
-    /**
-     * A username to be used with XMLHttpRequest in response to an HTTP access authentication request.
-     */
-    username?: string;
-    /**
-     * Callback for creating the XMLHttpRequest object. Defaults to the ActiveXObject when available (IE), the XMLHttpRequest otherwise. Override to provide your own implementation for XMLHttpRequest or enhancements to the factory.
-     */
-    xhr?: any;
-    /**
-     * An object of fieldName-fieldValue pairs to set on the native XHR object. For example, you can use it to set withCredentials to true for cross-domain requests if needed. In jQuery 1.5, the withCredentials property was not propagated to the native XHR and thus CORS requests requiring it would ignore this flag. For this reason, we recommend using jQuery 1.5.1+ should you require the use of it. (version added: 1.5.1)
-     */
-    xhrFields?: { [key: string]: any; };
-}
-
-/**
- * Interface for the jqXHR object
- */
-interface JQueryXHR extends XMLHttpRequest, JQueryPromise<any> {
-    /**
-     * The .overrideMimeType() method may be used in the beforeSend() callback function, for example, to modify the response content-type header. As of jQuery 1.5.1, the jqXHR object also contains the overrideMimeType() method (it was available in jQuery 1.4.x, as well, but was temporarily removed in jQuery 1.5). 
-     */
-    overrideMimeType(mimeType: string): any;
-    /**
-     * Cancel the request. 
-     *
-     * @param statusText A string passed as the textStatus parameter for the done callback. Default value: "canceled"
-     */
-    abort(statusText?: string): void;
-    /**
-     * Incorporates the functionality of the .done() and .fail() methods, allowing (as of jQuery 1.8) the underlying Promise to be manipulated. Refer to deferred.then() for implementation details.
-     */
-    then<R>(doneCallback: (data: any, textStatus: string, jqXHR: JQueryXHR) => R, failCallback?: (jqXHR: JQueryXHR, textStatus: string, errorThrown: any) => void): JQueryPromise<R>;
-    /**
-     * Property containing the parsed response if the response Content-Type is json
-     */
-    responseJSON?: any;
-    /**
-     * A function to be called if the request fails.
-     */
-    error(xhr: JQueryXHR, textStatus: string, errorThrown: string): void;
-}
-
-/**
- * Interface for the JQuery callback
- */
-interface JQueryCallback {
-    /**
-     * Add a callback or a collection of callbacks to a callback list.
-     * 
-     * @param callbacks A function, or array of functions, that are to be added to the callback list.
-     */
-    add(callbacks: Function): JQueryCallback;
-    /**
-     * Add a callback or a collection of callbacks to a callback list.
-     * 
-     * @param callbacks A function, or array of functions, that are to be added to the callback list.
-     */
-    add(callbacks: Function[]): JQueryCallback;
-
-    /**
-     * Disable a callback list from doing anything more.
-     */
-    disable(): JQueryCallback;
-
-    /**
-     * Determine if the callbacks list has been disabled.
-     */
-    disabled(): boolean;
-
-    /**
-     * Remove all of the callbacks from a list.
-     */
-    empty(): JQueryCallback;
-
-    /**
-     * Call all of the callbacks with the given arguments
-     * 
-     * @param arguments The argument or list of arguments to pass back to the callback list.
-     */
-    fire(...arguments: any[]): JQueryCallback;
-
-    /**
-     * Determine if the callbacks have already been called at least once.
-     */
-    fired(): boolean;
-
-    /**
-     * Call all callbacks in a list with the given context and arguments.
-     * 
-     * @param context A reference to the context in which the callbacks in the list should be fired.
-     * @param arguments An argument, or array of arguments, to pass to the callbacks in the list.
-     */
-    fireWith(context?: any, args?: any[]): JQueryCallback;
-
-    /**
-     * Determine whether a supplied callback is in a list
-     * 
-     * @param callback The callback to search for.
-     */
-    has(callback: Function): boolean;
-
-    /**
-     * Lock a callback list in its current state.
-     */
-    lock(): JQueryCallback;
-
-    /**
-     * Determine if the callbacks list has been locked.
-     */
-    locked(): boolean;
-
-    /**
-     * Remove a callback or a collection of callbacks from a callback list.
-     * 
-     * @param callbacks A function, or array of functions, that are to be removed from the callback list.
-     */
-    remove(callbacks: Function): JQueryCallback;
-    /**
-     * Remove a callback or a collection of callbacks from a callback list.
-     * 
-     * @param callbacks A function, or array of functions, that are to be removed from the callback list.
-     */
-    remove(callbacks: Function[]): JQueryCallback;
-}
-
-/**
- * Allows jQuery Promises to interop with non-jQuery promises
- */
-interface JQueryGenericPromise<T> {
-    /**
-     * Add handlers to be called when the Deferred object is resolved, rejected, or still in progress.
-     * 
-     * @param doneFilter A function that is called when the Deferred is resolved.
-     * @param failFilter An optional function that is called when the Deferred is rejected.
-     */
-    then<U>(doneFilter: (value?: T, ...values: any[]) => U|JQueryPromise<U>, failFilter?: (...reasons: any[]) => any, progressFilter?: (...progression: any[]) => any): JQueryPromise<U>;
-
-    /**
-     * Add handlers to be called when the Deferred object is resolved, rejected, or still in progress.
-     * 
-     * @param doneFilter A function that is called when the Deferred is resolved.
-     * @param failFilter An optional function that is called when the Deferred is rejected.
-     */
-    then(doneFilter: (value?: T, ...values: any[]) => void, failFilter?: (...reasons: any[]) => any, progressFilter?: (...progression: any[]) => any): JQueryPromise<void>;
-}
-
-/**
- * Interface for the JQuery promise/deferred callbacks
- */
-interface JQueryPromiseCallback<T> {
-    (value?: T, ...args: any[]): void;
-}
-
-interface JQueryPromiseOperator<T, U> {
-    (callback1: JQueryPromiseCallback<T>|JQueryPromiseCallback<T>[], ...callbacksN: Array<JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[]>): JQueryPromise<U>;
-}
-
-/**
- * Interface for the JQuery promise, part of callbacks
- */
-interface JQueryPromise<T> extends JQueryGenericPromise<T> {
-    /**
-     * Determine the current state of a Deferred object.
-     */
-    state(): string;
-    /**
-     * Add handlers to be called when the Deferred object is either resolved or rejected.
-     * 
-     * @param alwaysCallbacks1 A function, or array of functions, that is called when the Deferred is resolved or rejected.
-     * @param alwaysCallbacks2 Optional additional functions, or arrays of functions, that are called when the Deferred is resolved or rejected.
-     */
-    always(alwaysCallback1?: JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[], ...alwaysCallbacksN: Array<JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[]>): JQueryPromise<T>;
-    /**
-     * Add handlers to be called when the Deferred object is resolved.
-     * 
-     * @param doneCallbacks1 A function, or array of functions, that are called when the Deferred is resolved.
-     * @param doneCallbacks2 Optional additional functions, or arrays of functions, that are called when the Deferred is resolved.
-     */
-    done(doneCallback1?: JQueryPromiseCallback<T>|JQueryPromiseCallback<T>[], ...doneCallbackN: Array<JQueryPromiseCallback<T>|JQueryPromiseCallback<T>[]>): JQueryPromise<T>;
-    /**
-     * Add handlers to be called when the Deferred object is rejected.
-     * 
-     * @param failCallbacks1 A function, or array of functions, that are called when the Deferred is rejected.
-     * @param failCallbacks2 Optional additional functions, or arrays of functions, that are called when the Deferred is rejected.
-     */
-    fail(failCallback1?: JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[], ...failCallbacksN: Array<JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[]>): JQueryPromise<T>;
-    /**
-     * Add handlers to be called when the Deferred object generates progress notifications.
-     * 
-     * @param progressCallbacks A function, or array of functions, to be called when the Deferred generates progress notifications.
-     */
-    progress(progressCallback1?: JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[], ...progressCallbackN: Array<JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[]>): JQueryPromise<T>;
-
-    // Deprecated - given no typings
-    pipe(doneFilter?: (x: any) => any, failFilter?: (x: any) => any, progressFilter?: (x: any) => any): JQueryPromise<any>;
-}
-
-/**
- * Interface for the JQuery deferred, part of callbacks
- */
-interface JQueryDeferred<T> extends JQueryGenericPromise<T> {
-    /**
-     * Determine the current state of a Deferred object.
-     */
-    state(): string;
-    /**
-     * Add handlers to be called when the Deferred object is either resolved or rejected.
-     * 
-     * @param alwaysCallbacks1 A function, or array of functions, that is called when the Deferred is resolved or rejected.
-     * @param alwaysCallbacks2 Optional additional functions, or arrays of functions, that are called when the Deferred is resolved or rejected.
-     */
-    always(alwaysCallback1?: JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[], ...alwaysCallbacksN: Array<JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[]>): JQueryDeferred<T>;
-    /**
-     * Add handlers to be called when the Deferred object is resolved.
-     * 
-     * @param doneCallbacks1 A function, or array of functions, that are called when the Deferred is resolved.
-     * @param doneCallbacks2 Optional additional functions, or arrays of functions, that are called when the Deferred is resolved.
-     */
-    done(doneCallback1?: JQueryPromiseCallback<T>|JQueryPromiseCallback<T>[], ...doneCallbackN: Array<JQueryPromiseCallback<T>|JQueryPromiseCallback<T>[]>): JQueryDeferred<T>;
-    /**
-     * Add handlers to be called when the Deferred object is rejected.
-     * 
-     * @param failCallbacks1 A function, or array of functions, that are called when the Deferred is rejected.
-     * @param failCallbacks2 Optional additional functions, or arrays of functions, that are called when the Deferred is rejected.
-     */
-    fail(failCallback1?: JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[], ...failCallbacksN: Array<JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[]>): JQueryDeferred<T>;
-    /**
-     * Add handlers to be called when the Deferred object generates progress notifications.
-     * 
-     * @param progressCallbacks A function, or array of functions, to be called when the Deferred generates progress notifications.
-     */
-    progress(progressCallback1?: JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[], ...progressCallbackN: Array<JQueryPromiseCallback<any>|JQueryPromiseCallback<any>[]>): JQueryDeferred<T>;
-
-    /**
-     * Call the progressCallbacks on a Deferred object with the given args.
-     * 
-     * @param args Optional arguments that are passed to the progressCallbacks.
-     */
-    notify(value?: any, ...args: any[]): JQueryDeferred<T>;
-
-    /**
-     * Call the progressCallbacks on a Deferred object with the given context and args.
-     * 
-     * @param context Context passed to the progressCallbacks as the this object.
-     * @param args Optional arguments that are passed to the progressCallbacks.
-     */
-    notifyWith(context: any, value?: any[]): JQueryDeferred<T>;
-
-    /**
-     * Reject a Deferred object and call any failCallbacks with the given args.
-     * 
-     * @param args Optional arguments that are passed to the failCallbacks.
-     */
-    reject(value?: any, ...args: any[]): JQueryDeferred<T>;
-    /**
-     * Reject a Deferred object and call any failCallbacks with the given context and args.
-     * 
-     * @param context Context passed to the failCallbacks as the this object.
-     * @param args An optional array of arguments that are passed to the failCallbacks.
-     */
-    rejectWith(context: any, value?: any[]): JQueryDeferred<T>;
-
-    /**
-     * Resolve a Deferred object and call any doneCallbacks with the given args.
-     * 
-     * @param value First argument passed to doneCallbacks.
-     * @param args Optional subsequent arguments that are passed to the doneCallbacks.
-     */
-    resolve(value?: T, ...args: any[]): JQueryDeferred<T>;
-
-    /**
-     * Resolve a Deferred object and call any doneCallbacks with the given context and args.
-     * 
-     * @param context Context passed to the doneCallbacks as the this object.
-     * @param args An optional array of arguments that are passed to the doneCallbacks.
-     */
-    resolveWith(context: any, value?: T[]): JQueryDeferred<T>;
-
-    /**
-     * Return a Deferred's Promise object.
-     * 
-     * @param target Object onto which the promise methods have to be attached
-     */
-    promise(target?: any): JQueryPromise<T>;
-
-    // Deprecated - given no typings
-    pipe(doneFilter?: (x: any) => any, failFilter?: (x: any) => any, progressFilter?: (x: any) => any): JQueryPromise<any>;
-}
-
-/**
- * Interface of the JQuery extension of the W3C event object
- */
-interface BaseJQueryEventObject extends Event {
-    data: any;
-    delegateTarget: Element;
-    isDefaultPrevented(): boolean;
-    isImmediatePropagationStopped(): boolean;
-    isPropagationStopped(): boolean;
-    namespace: string;
-    originalEvent: Event;
-    preventDefault(): any;
-    relatedTarget: Element;
-    result: any;
-    stopImmediatePropagation(): void;
-    stopPropagation(): void;
-    target: Element;
-    pageX: number;
-    pageY: number;
-    which: number;
-    metaKey: boolean;
-}
-
-interface JQueryInputEventObject extends BaseJQueryEventObject {
-    altKey: boolean;
-    ctrlKey: boolean;
-    metaKey: boolean;
-    shiftKey: boolean;
-}
-
-interface JQueryMouseEventObject extends JQueryInputEventObject {
-    button: number;
-    clientX: number;
-    clientY: number;
-    offsetX: number;
-    offsetY: number;
-    pageX: number;
-    pageY: number;
-    screenX: number;
-    screenY: number;
-}
-
-interface JQueryKeyEventObject extends JQueryInputEventObject {
-    char: any;
-    charCode: number;
-    key: any;
-    keyCode: number;
-}
-
-interface JQueryEventObject extends BaseJQueryEventObject, JQueryInputEventObject, JQueryMouseEventObject, JQueryKeyEventObject{
-}
-
-/*
-    Collection of properties of the current browser
-*/
-
-interface JQuerySupport {
-    ajax?: boolean;
-    boxModel?: boolean;
-    changeBubbles?: boolean;
-    checkClone?: boolean;
-    checkOn?: boolean;
-    cors?: boolean;
-    cssFloat?: boolean;
-    hrefNormalized?: boolean;
-    htmlSerialize?: boolean;
-    leadingWhitespace?: boolean;
-    noCloneChecked?: boolean;
-    noCloneEvent?: boolean;
-    opacity?: boolean;
-    optDisabled?: boolean;
-    optSelected?: boolean;
-    scriptEval? (): boolean;
-    style?: boolean;
-    submitBubbles?: boolean;
-    tbody?: boolean;
-}
-
-interface JQueryParam {
-    /**
-     * Create a serialized representation of an array or object, suitable for use in a URL query string or Ajax request.
-     * 
-     * @param obj An array or object to serialize.
-     */
-    (obj: any): string;
-
-    /**
-     * Create a serialized representation of an array or object, suitable for use in a URL query string or Ajax request.
-     * 
-     * @param obj An array or object to serialize.
-     * @param traditional A Boolean indicating whether to perform a traditional "shallow" serialization.
-     */
-    (obj: any, traditional: boolean): string;
-}
-
-/**
- * The interface used to construct jQuery events (with $.Event). It is
- * defined separately instead of inline in JQueryStatic to allow
- * overriding the construction function with specific strings
- * returning specific event objects.
- */
-interface JQueryEventConstructor {
-    (name: string, eventProperties?: any): JQueryEventObject;
-    new (name: string, eventProperties?: any): JQueryEventObject;
-}
-
-/**
- * The interface used to specify coordinates.
- */
-interface JQueryCoordinates {
-    left: number;
-    top: number;
-}
-
-/**
- * Elements in the array returned by serializeArray()
- */
-interface JQuerySerializeArrayElement {
-    name: string;
-    value: string;
-}
-
-interface JQueryAnimationOptions { 
-    /**
-     * A string or number determining how long the animation will run.
-     */
-    duration?: any; 
-    /**
-     * A string indicating which easing function to use for the transition.
-     */
-    easing?: string; 
-    /**
-     * A function to call once the animation is complete.
-     */
-    complete?: Function; 
-    /**
-     * A function to be called for each animated property of each animated element. This function provides an opportunity to modify the Tween object to change the value of the property before it is set.
-     */
-    step?: (now: number, tween: any) => any; 
-    /**
-     * A function to be called after each step of the animation, only once per animated element regardless of the number of animated properties. (version added: 1.8)
-     */
-    progress?: (animation: JQueryPromise<any>, progress: number, remainingMs: number) => any; 
-    /**
-     * A function to call when the animation begins. (version added: 1.8)
-     */
-    start?: (animation: JQueryPromise<any>) => any; 
-    /**
-     * A function to be called when the animation completes (its Promise object is resolved). (version added: 1.8)
-     */
-    done?: (animation: JQueryPromise<any>, jumpedToEnd: boolean) => any; 
-    /**
-     * A function to be called when the animation fails to complete (its Promise object is rejected). (version added: 1.8)
-     */
-    fail?: (animation: JQueryPromise<any>, jumpedToEnd: boolean) => any; 
-    /**
-     * A function to be called when the animation completes or stops without completing (its Promise object is either resolved or rejected). (version added: 1.8)
-     */
-    always?: (animation: JQueryPromise<any>, jumpedToEnd: boolean) => any; 
-    /**
-     * A Boolean indicating whether to place the animation in the effects queue. If false, the animation will begin immediately. As of jQuery 1.7, the queue option can also accept a string, in which case the animation is added to the queue represented by that string. When a custom queue name is used the animation does not automatically start; you must call .dequeue("queuename") to start it.
-     */
-    queue?: any; 
-    /**
-     * A map of one or more of the CSS properties defined by the properties argument and their corresponding easing functions. (version added: 1.4)
-     */
-    specialEasing?: Object;
-}
-
-interface JQueryEasingFunction {
-    ( percent: number ): number;
-}
-
-interface JQueryEasingFunctions {
-    [ name: string ]: JQueryEasingFunction;
-    linear: JQueryEasingFunction;
-    swing: JQueryEasingFunction;
-}
-
-/**
- * Static members of jQuery (those on $ and jQuery themselves)
- */
-interface JQueryStatic {
-
-    /**
-     * Perform an asynchronous HTTP (Ajax) request.
-     *
-     * @param settings A set of key/value pairs that configure the Ajax request. All settings are optional. A default can be set for any option with $.ajaxSetup().
-     */
-    ajax(settings: JQueryAjaxSettings): JQueryXHR;
-    /**
-     * Perform an asynchronous HTTP (Ajax) request.
-     *
-     * @param url A string containing the URL to which the request is sent.
-     * @param settings A set of key/value pairs that configure the Ajax request. All settings are optional. A default can be set for any option with $.ajaxSetup().
-     */
-    ajax(url: string, settings?: JQueryAjaxSettings): JQueryXHR;
-
-    /**
-     * Handle custom Ajax options or modify existing options before each request is sent and before they are processed by $.ajax().
-     *
-     * @param dataTypes An optional string containing one or more space-separated dataTypes
-     * @param handler A handler to set default values for future Ajax requests.
-     */
-    ajaxPrefilter(dataTypes: string, handler: (opts: any, originalOpts: JQueryAjaxSettings, jqXHR: JQueryXHR) => any): void;
-    /**
-     * Handle custom Ajax options or modify existing options before each request is sent and before they are processed by $.ajax().
-     *
-     * @param handler A handler to set default values for future Ajax requests.
-     */
-    ajaxPrefilter(handler: (opts: any, originalOpts: JQueryAjaxSettings, jqXHR: JQueryXHR) => any): void;
-
-    ajaxSettings: JQueryAjaxSettings;
-
-     /**
-      * Set default values for future Ajax requests. Its use is not recommended.
-      *
-      * @param options A set of key/value pairs that configure the default Ajax request. All options are optional.
-      */
-    ajaxSetup(options: JQueryAjaxSettings): void;
-
-    /**
-     * Load data from the server using a HTTP GET request.
-     *
-     * @param url A string containing the URL to which the request is sent.
-     * @param success A callback function that is executed if the request succeeds.
-     * @param dataType The type of data expected from the server. Default: Intelligent Guess (xml, json, script, or html).
-     */
-    get(url: string, success?: (data: any, textStatus: string, jqXHR: JQueryXHR) => any, dataType?: string): JQueryXHR;
-    /**
-     * Load data from the server using a HTTP GET request.
-     *
-     * @param url A string containing the URL to which the request is sent.
-     * @param data A plain object or string that is sent to the server with the request.
-     * @param success A callback function that is executed if the request succeeds.
-     * @param dataType The type of data expected from the server. Default: Intelligent Guess (xml, json, script, or html).
-     */
-    get(url: string, data?: Object|string, success?: (data: any, textStatus: string, jqXHR: JQueryXHR) => any, dataType?: string): JQueryXHR;
-    /**
-     * Load data from the server using a HTTP GET request.
-     *
-     * @param settings The JQueryAjaxSettings to be used for the request
-     */
-    get(settings : JQueryAjaxSettings): JQueryXHR;
-    /**
-     * Load JSON-encoded data from the server using a GET HTTP request.
-     *
-     * @param url A string containing the URL to which the request is sent.
-     * @param success A callback function that is executed if the request succeeds.
-     */
-    getJSON(url: string, success?: (data: any, textStatus: string, jqXHR: JQueryXHR) => any): JQueryXHR;
-    /**
-     * Load JSON-encoded data from the server using a GET HTTP request.
-     *
-     * @param url A string containing the URL to which the request is sent.
-     * @param data A plain object or string that is sent to the server with the request.
-     * @param success A callback function that is executed if the request succeeds.
-     */
-    getJSON(url: string, data?: Object|string, success?: (data: any, textStatus: string, jqXHR: JQueryXHR) => any): JQueryXHR;
-    /**
-     * Load a JavaScript file from the server using a GET HTTP request, then execute it.
-     *
-     * @param url A string containing the URL to which the request is sent.
-     * @param success A callback function that is executed if the request succeeds.
-     */
-    getScript(url: string, success?: (script: string, textStatus: string, jqXHR: JQueryXHR) => any): JQueryXHR;
-
-    /**
-     * Create a serialized representation of an array or object, suitable for use in a URL query string or Ajax request.
-     */
-    param: JQueryParam;
-
-    /**
-     * Load data from the server using a HTTP POST request.
-     *
-     * @param url A string containing the URL to which the request is sent.
-     * @param success A callback function that is executed if the request succeeds. Required if dataType is provided, but can be null in that case.
-     * @param dataType The type of data expected from the server. Default: Intelligent Guess (xml, json, script, text, html).
-     */
-    post(url: string, success?: (data: any, textStatus: string, jqXHR: JQueryXHR) => any, dataType?: string): JQueryXHR;
-    /**
-     * Load data from the server using a HTTP POST request.
-     *
-     * @param url A string containing the URL to which the request is sent.
-     * @param data A plain object or string that is sent to the server with the request.
-     * @param success A callback function that is executed if the request succeeds. Required if dataType is provided, but can be null in that case.
-     * @param dataType The type of data expected from the server. Default: Intelligent Guess (xml, json, script, text, html).
-     */
-    post(url: string, data?: Object|string, success?: (data: any, textStatus: string, jqXHR: JQueryXHR) => any, dataType?: string): JQueryXHR;
-    /**
-     * Load data from the server using a HTTP POST request.
-     *
-     * @param settings The JQueryAjaxSettings to be used for the request
-     */
-    post(settings : JQueryAjaxSettings): JQueryXHR;
-    /**
-     * A multi-purpose callbacks list object that provides a powerful way to manage callback lists.
-     *
-     * @param flags An optional list of space-separated flags that change how the callback list behaves.
-     */
-    Callbacks(flags?: string): JQueryCallback;
-
-    /**
-     * Holds or releases the execution of jQuery's ready event.
-     *
-     * @param hold Indicates whether the ready hold is being requested or released
-     */
-    holdReady(hold: boolean): void;
-
-    /**
-     * Accepts a string containing a CSS selector which is then used to match a set of elements.
-     *
-     * @param selector A string containing a selector expression
-     * @param context A DOM Element, Document, or jQuery to use as context
-     */
-    (selector: string, context?: Element|JQuery): JQuery;
-
-    /**
-     * Accepts a string containing a CSS selector which is then used to match a set of elements.
-     *
-     * @param element A DOM element to wrap in a jQuery object.
-     */
-    (element: Element): JQuery;
-
-    /**
-     * Accepts a string containing a CSS selector which is then used to match a set of elements.
-     *
-     * @param elementArray An array containing a set of DOM elements to wrap in a jQuery object.
-     */
-    (elementArray: Element[]): JQuery;
-
-    /**
-     * Binds a function to be executed when the DOM has finished loading.
-     *
-     * @param callback A function to execute after the DOM is ready.
-     */
-    (callback: (jQueryAlias?: JQueryStatic) => any): JQuery;
-
-    /**
-     * Accepts a string containing a CSS selector which is then used to match a set of elements.
-     *
-     * @param object A plain object to wrap in a jQuery object.
-     */
-    (object: {}): JQuery;
-
-    /**
-     * Accepts a string containing a CSS selector which is then used to match a set of elements.
-     *
-     * @param object An existing jQuery object to clone.
-     */
-    (object: JQuery): JQuery;
-
-    /**
-     * Specify a function to execute when the DOM is fully loaded.
-     */
-    (): JQuery;
-
-    /**
-     * Creates DOM elements on the fly from the provided string of raw HTML.
-     *
-     * @param html A string of HTML to create on the fly. Note that this parses HTML, not XML.
-     * @param ownerDocument A document in which the new elements will be created.
-     */
-    (html: string, ownerDocument?: Document): JQuery;
-
-    /**
-     * Creates DOM elements on the fly from the provided string of raw HTML.
-     *
-     * @param html A string defining a single, standalone, HTML element (e.g. <div/> or <div></div>).
-     * @param attributes An object of attributes, events, and methods to call on the newly-created element.
-     */
-    (html: string, attributes: Object): JQuery;
-
-    /**
-     * Relinquish jQuery's control of the $ variable.
-     *
-     * @param removeAll A Boolean indicating whether to remove all jQuery variables from the global scope (including jQuery itself).
-     */
-    noConflict(removeAll?: boolean): JQueryStatic;
-
-    /**
-     * Provides a way to execute callback functions based on one or more objects, usually Deferred objects that represent asynchronous events.
-     *
-     * @param deferreds One or more Deferred objects, or plain JavaScript objects.
-     */
-    when<T>(...deferreds: Array<T|JQueryPromise<T>/* as JQueryDeferred<T> */>): JQueryPromise<T>;
-
-    /**
-     * Hook directly into jQuery to override how particular CSS properties are retrieved or set, normalize CSS property naming, or create custom properties.
-     */
-    cssHooks: { [key: string]: any; };
-    cssNumber: any;
-
-    /**
-     * Store arbitrary data associated with the specified element. Returns the value that was set.
-     *
-     * @param element The DOM element to associate with the data.
-     * @param key A string naming the piece of data to set.
-     * @param value The new data value.
-     */
-    data<T>(element: Element, key: string, value: T): T;
-    /**
-     * Returns value at named data store for the element, as set by jQuery.data(element, name, value), or the full data store for the element.
-     *
-     * @param element The DOM element to associate with the data.
-     * @param key A string naming the piece of data to set.
-     */
-    data(element: Element, key: string): any;
-    /**
-     * Returns value at named data store for the element, as set by jQuery.data(element, name, value), or the full data store for the element.
-     *
-     * @param element The DOM element to associate with the data.
-     */
-    data(element: Element): any;
-
-    /**
-     * Execute the next function on the queue for the matched element.
-     *
-     * @param element A DOM element from which to remove and execute a queued function.
-     * @param queueName A string containing the name of the queue. Defaults to fx, the standard effects queue.
-     */
-    dequeue(element: Element, queueName?: string): void;
-
-    /**
-     * Determine whether an element has any jQuery data associated with it.
-     *
-     * @param element A DOM element to be checked for data.
-     */
-    hasData(element: Element): boolean;
-
-    /**
-     * Show the queue of functions to be executed on the matched element.
-     *
-     * @param element A DOM element to inspect for an attached queue.
-     * @param queueName A string containing the name of the queue. Defaults to fx, the standard effects queue.
-     */
-    queue(element: Element, queueName?: string): any[];
-    /**
-     * Manipulate the queue of functions to be executed on the matched element.
-     *
-     * @param element A DOM element where the array of queued functions is attached.
-     * @param queueName A string containing the name of the queue. Defaults to fx, the standard effects queue.
-     * @param newQueue An array of functions to replace the current queue contents.
-     */
-    queue(element: Element, queueName: string, newQueue: Function[]): JQuery;
-    /**
-     * Manipulate the queue of functions to be executed on the matched element.
-     *
-     * @param element A DOM element on which to add a queued function.
-     * @param queueName A string containing the name of the queue. Defaults to fx, the standard effects queue.
-     * @param callback The new function to add to the queue.
-     */
-    queue(element: Element, queueName: string, callback: Function): JQuery;
-
-    /**
-     * Remove a previously-stored piece of data.
-     *
-     * @param element A DOM element from which to remove data.
-     * @param name A string naming the piece of data to remove.
-     */
-    removeData(element: Element, name?: string): JQuery;
-
-    /**
-     * A constructor function that returns a chainable utility object with methods to register multiple callbacks into callback queues, invoke callback queues, and relay the success or failure state of any synchronous or asynchronous function.
-     *
-     * @param beforeStart A function that is called just before the constructor returns.
-     */
-    Deferred<T>(beforeStart?: (deferred: JQueryDeferred<T>) => any): JQueryDeferred<T>;
-
-    /**
-     * Effects
-     */
-
-    easing: JQueryEasingFunctions;
-
-    fx: {
-        tick: () => void;
-        /**
-         * The rate (in milliseconds) at which animations fire.
-         */
-        interval: number;
-        stop: () => void;
-        speeds: { slow: number; fast: number; };
-        /**
-         * Globally disable all animations.
-         */
-        off: boolean;
-        step: any;
-    };
-
-    /**
-     * Takes a function and returns a new one that will always have a particular context.
-     *
-     * @param fnction The function whose context will be changed.
-     * @param context The object to which the context (this) of the function should be set.
-     * @param additionalArguments Any number of arguments to be passed to the function referenced in the function argument.
-     */
-    proxy(fnction: (...args: any[]) => any, context: Object, ...additionalArguments: any[]): any;
-    /**
-     * Takes a function and returns a new one that will always have a particular context.
-     *
-     * @param context The object to which the context (this) of the function should be set.
-     * @param name The name of the function whose context will be changed (should be a property of the context object).
-     * @param additionalArguments Any number of arguments to be passed to the function named in the name argument.
-     */
-    proxy(context: Object, name: string, ...additionalArguments: any[]): any;
-
-    Event: JQueryEventConstructor;
-
-    /**
-     * Takes a string and throws an exception containing it.
-     *
-     * @param message The message to send out.
-     */
-    error(message: any): JQuery;
-
-    expr: any;
-    fn: any;  //TODO: Decide how we want to type this
-
-    isReady: boolean;
-
-    // Properties
-    support: JQuerySupport;
-
-    /**
-     * Check to see if a DOM element is a descendant of another DOM element.
-     * 
-     * @param container The DOM element that may contain the other element.
-     * @param contained The DOM element that may be contained by (a descendant of) the other element.
-     */
-    contains(container: Element, contained: Element): boolean;
-
-    /**
-     * A generic iterator function, which can be used to seamlessly iterate over both objects and arrays. Arrays and array-like objects with a length property (such as a function's arguments object) are iterated by numeric index, from 0 to length-1. Other objects are iterated via their named properties.
-     * 
-     * @param collection The object or array to iterate over.
-     * @param callback The function that will be executed on every object.
-     */
-    each<T>(
-        collection: T[],
-        callback: (indexInArray: number, valueOfElement: T) => any
-        ): any;
-
-    /**
-     * A generic iterator function, which can be used to seamlessly iterate over both objects and arrays. Arrays and array-like objects with a length property (such as a function's arguments object) are iterated by numeric index, from 0 to length-1. Other objects are iterated via their named properties.
-     * 
-     * @param collection The object or array to iterate over.
-     * @param callback The function that will be executed on every object.
-     */
-    each(
-        collection: any,
-        callback: (indexInArray: any, valueOfElement: any) => any
-        ): any;
-
-    /**
-     * Merge the contents of two or more objects together into the first object.
-     *
-     * @param target An object that will receive the new properties if additional objects are passed in or that will extend the jQuery namespace if it is the sole argument.
-     * @param object1 An object containing additional properties to merge in.
-     * @param objectN Additional objects containing properties to merge in.
-     */
-    extend(target: any, object1?: any, ...objectN: any[]): any;
-    /**
-     * Merge the contents of two or more objects together into the first object.
-     *
-     * @param deep If true, the merge becomes recursive (aka. deep copy).
-     * @param target The object to extend. It will receive the new properties.
-     * @param object1 An object containing additional properties to merge in.
-     * @param objectN Additional objects containing properties to merge in.
-     */
-    extend(deep: boolean, target: any, object1?: any, ...objectN: any[]): any;
-
-    /**
-     * Execute some JavaScript code globally.
-     *
-     * @param code The JavaScript code to execute.
-     */
-    globalEval(code: string): any;
-
-    /**
-     * Finds the elements of an array which satisfy a filter function. The original array is not affected.
-     *
-     * @param array The array to search through.
-     * @param func The function to process each item against. The first argument to the function is the item, and the second argument is the index. The function should return a Boolean value.  this will be the global window object.
-     * @param invert If "invert" is false, or not provided, then the function returns an array consisting of all elements for which "callback" returns true. If "invert" is true, then the function returns an array consisting of all elements for which "callback" returns false.
-     */
-    grep<T>(array: T[], func: (elementOfArray: T, indexInArray: number) => boolean, invert?: boolean): T[];
-
-    /**
-     * Search for a specified value within an array and return its index (or -1 if not found).
-     *
-     * @param value The value to search for.
-     * @param array An array through which to search.
-     * @param fromIndex he index of the array at which to begin the search. The default is 0, which will search the whole array.
-     */
-    inArray<T>(value: T, array: T[], fromIndex?: number): number;
-
-    /**
-     * Determine whether the argument is an array.
-     *
-     * @param obj Object to test whether or not it is an array.
-     */
-    isArray(obj: any): boolean;
-    /**
-     * Check to see if an object is empty (contains no enumerable properties).
-     *
-     * @param obj The object that will be checked to see if it's empty.
-     */
-    isEmptyObject(obj: any): boolean;
-    /**
-     * Determine if the argument passed is a Javascript function object.
-     *
-     * @param obj Object to test whether or not it is a function.
-     */
-    isFunction(obj: any): boolean;
-    /**
-     * Determines whether its argument is a number.
-     *
-     * @param obj The value to be tested.
-     */
-    isNumeric(value: any): boolean;
-    /**
-     * Check to see if an object is a plain object (created using "{}" or "new Object").
-     *
-     * @param obj The object that will be checked to see if it's a plain object.
-     */
-    isPlainObject(obj: any): boolean;
-    /**
-     * Determine whether the argument is a window.
-     *
-     * @param obj Object to test whether or not it is a window.
-     */
-    isWindow(obj: any): boolean;
-    /**
-     * Check to see if a DOM node is within an XML document (or is an XML document).
-     *
-     * @param node he DOM node that will be checked to see if it's in an XML document.
-     */
-    isXMLDoc(node: Node): boolean;
-
-    /**
-     * Convert an array-like object into a true JavaScript array.
-     * 
-     * @param obj Any object to turn into a native Array.
-     */
-    makeArray(obj: any): any[];
-
-    /**
-     * Translate all items in an array or object to new array of items.
-     * 
-     * @param array The Array to translate.
-     * @param callback The function to process each item against. The first argument to the function is the array item, the second argument is the index in array The function can return any value. Within the function, this refers to the global (window) object.
-     */
-    map<T, U>(array: T[], callback: (elementOfArray: T, indexInArray: number) => U): U[];
-    /**
-     * Translate all items in an array or object to new array of items.
-     * 
-     * @param arrayOrObject The Array or Object to translate.
-     * @param callback The function to process each item against. The first argument to the function is the value; the second argument is the index or key of the array or object property. The function can return any value to add to the array. A returned array will be flattened into the resulting array. Within the function, this refers to the global (window) object.
-     */
-    map(arrayOrObject: any, callback: (value: any, indexOrKey: any) => any): any;
-
-    /**
-     * Merge the contents of two arrays together into the first array.
-     * 
-     * @param first The first array to merge, the elements of second added.
-     * @param second The second array to merge into the first, unaltered.
-     */
-    merge<T>(first: T[], second: T[]): T[];
-
-    /**
-     * An empty function.
-     */
-    noop(): any;
-
-    /**
-     * Return a number representing the current time.
-     */
-    now(): number;
-
-    /**
-     * Takes a well-formed JSON string and returns the resulting JavaScript object.
-     * 
-     * @param json The JSON string to parse.
-     */
-    parseJSON(json: string): any;
-
-    /**
-     * Parses a string into an XML document.
-     *
-     * @param data a well-formed XML string to be parsed
-     */
-    parseXML(data: string): XMLDocument;
-
-    /**
-     * Remove the whitespace from the beginning and end of a string.
-     * 
-     * @param str Remove the whitespace from the beginning and end of a string.
-     */
-    trim(str: string): string;
-
-    /**
-     * Determine the internal JavaScript [[Class]] of an object.
-     * 
-     * @param obj Object to get the internal JavaScript [[Class]] of.
-     */
-    type(obj: any): string;
-
-    /**
-     * Sorts an array of DOM elements, in place, with the duplicates removed. Note that this only works on arrays of DOM elements, not strings or numbers.
-     * 
-     * @param array The Array of DOM elements.
-     */
-    unique(array: Element[]): Element[];
-
-    /**
-     * Parses a string into an array of DOM nodes.
-     *
-     * @param data HTML string to be parsed
-     * @param context DOM element to serve as the context in which the HTML fragment will be created
-     * @param keepScripts A Boolean indicating whether to include scripts passed in the HTML string
-     */
-    parseHTML(data: string, context?: HTMLElement, keepScripts?: boolean): any[];
-
-    /**
-     * Parses a string into an array of DOM nodes.
-     *
-     * @param data HTML string to be parsed
-     * @param context DOM element to serve as the context in which the HTML fragment will be created
-     * @param keepScripts A Boolean indicating whether to include scripts passed in the HTML string
-     */
-    parseHTML(data: string, context?: Document, keepScripts?: boolean): any[];
-}
-
-/**
- * The jQuery instance members
- */
-interface JQuery {
-    /**
-     * Register a handler to be called when Ajax requests complete. This is an AjaxEvent.
-     *
-     * @param handler The function to be invoked.
-     */
-    ajaxComplete(handler: (event: JQueryEventObject, XMLHttpRequest: XMLHttpRequest, ajaxOptions: any) => any): JQuery;
-    /**
-     * Register a handler to be called when Ajax requests complete with an error. This is an Ajax Event.
-     *
-     * @param handler The function to be invoked.
-     */
-    ajaxError(handler: (event: JQueryEventObject, jqXHR: JQueryXHR, ajaxSettings: JQueryAjaxSettings, thrownError: any) => any): JQuery;
-    /**
-     * Attach a function to be executed before an Ajax request is sent. This is an Ajax Event.
-     *
-     * @param handler The function to be invoked.
-     */
-    ajaxSend(handler: (event: JQueryEventObject, jqXHR: JQueryXHR, ajaxOptions: JQueryAjaxSettings) => any): JQuery;
-    /**
-     * Register a handler to be called when the first Ajax request begins. This is an Ajax Event.
-     *
-     * @param handler The function to be invoked.
-     */
-    ajaxStart(handler: () => any): JQuery;
-    /**
-     * Register a handler to be called when all Ajax requests have completed. This is an Ajax Event.
-     *
-     * @param handler The function to be invoked.
-     */
-    ajaxStop(handler: () => any): JQuery;
-    /**
-     * Attach a function to be executed whenever an Ajax request completes successfully. This is an Ajax Event.
-     *
-     * @param handler The function to be invoked.
-     */
-    ajaxSuccess(handler: (event: JQueryEventObject, XMLHttpRequest: XMLHttpRequest, ajaxOptions: JQueryAjaxSettings) => any): JQuery;
-
-    /**
-     * Load data from the server and place the returned HTML into the matched element.
-     *
-     * @param url A string containing the URL to which the request is sent.
-     * @param data A plain object or string that is sent to the server with the request.
-     * @param complete A callback function that is executed when the request completes.
-     */
-    load(url: string, data?: string|Object, complete?: (responseText: string, textStatus: string, XMLHttpRequest: XMLHttpRequest) => any): JQuery;
-
-    /**
-     * Encode a set of form elements as a string for submission.
-     */
-    serialize(): string;
-    /**
-     * Encode a set of form elements as an array of names and values.
-     */
-    serializeArray(): JQuerySerializeArrayElement[];
-
-    /**
-     * Adds the specified class(es) to each of the set of matched elements.
-     *
-     * @param className One or more space-separated classes to be added to the class attribute of each matched element.
-     */
-    addClass(className: string): JQuery;
-    /**
-     * Adds the specified class(es) to each of the set of matched elements.
-     *
-     * @param function A function returning one or more space-separated class names to be added to the existing class name(s). Receives the index position of the element in the set and the existing class name(s) as arguments. Within the function, this refers to the current element in the set.
-     */
-    addClass(func: (index: number, className: string) => string): JQuery;
-
-    /**
-     * Add the previous set of elements on the stack to the current set, optionally filtered by a selector.
-     */
-    addBack(selector?: string): JQuery;
-
-    /**
-     * Get the value of an attribute for the first element in the set of matched elements.
-     *
-     * @param attributeName The name of the attribute to get.
-     */
-    attr(attributeName: string): string;
-    /**
-     * Set one or more attributes for the set of matched elements.
-     *
-     * @param attributeName The name of the attribute to set.
-     * @param value A value to set for the attribute.
-     */
-    attr(attributeName: string, value: string|number): JQuery;
-    /**
-     * Set one or more attributes for the set of matched elements.
-     *
-     * @param attributeName The name of the attribute to set.
-     * @param func A function returning the value to set. this is the current element. Receives the index position of the element in the set and the old attribute value as arguments.
-     */
-    attr(attributeName: string, func: (index: number, attr: string) => string|number): JQuery;
-    /**
-     * Set one or more attributes for the set of matched elements.
-     *
-     * @param attributes An object of attribute-value pairs to set.
-     */
-    attr(attributes: Object): JQuery;
-    
-    /**
-     * Determine whether any of the matched elements are assigned the given class.
-     *
-     * @param className The class name to search for.
-     */
-    hasClass(className: string): boolean;
-
-    /**
-     * Get the HTML contents of the first element in the set of matched elements.
-     */
-    html(): string;
-    /**
-     * Set the HTML contents of each element in the set of matched elements.
-     *
-     * @param htmlString A string of HTML to set as the content of each matched element.
-     */
-    html(htmlString: string): JQuery;
-    /**
-     * Set the HTML contents of each element in the set of matched elements.
-     *
-     * @param func A function returning the HTML content to set. Receives the index position of the element in the set and the old HTML value as arguments. jQuery empties the element before calling the function; use the oldhtml argument to reference the previous content. Within the function, this refers to the current element in the set.
-     */
-    html(func: (index: number, oldhtml: string) => string): JQuery;
-    /**
-     * Set the HTML contents of each element in the set of matched elements.
-     *
-     * @param func A function returning the HTML content to set. Receives the index position of the element in the set and the old HTML value as arguments. jQuery empties the element before calling the function; use the oldhtml argument to reference the previous content. Within the function, this refers to the current element in the set.
-     */
-
-    /**
-     * Get the value of a property for the first element in the set of matched elements.
-     *
-     * @param propertyName The name of the property to get.
-     */
-    prop(propertyName: string): any;
-    /**
-     * Set one or more properties for the set of matched elements.
-     *
-     * @param propertyName The name of the property to set.
-     * @param value A value to set for the property.
-     */
-    prop(propertyName: string, value: string|number|boolean): JQuery;
-    /**
-     * Set one or more properties for the set of matched elements.
-     *
-     * @param properties An object of property-value pairs to set.
-     */
-    prop(properties: Object): JQuery;
-    /**
-     * Set one or more properties for the set of matched elements.
-     *
-     * @param propertyName The name of the property to set.
-     * @param func A function returning the value to set. Receives the index position of the element in the set and the old property value as arguments. Within the function, the keyword this refers to the current element.
-     */
-    prop(propertyName: string, func: (index: number, oldPropertyValue: any) => any): JQuery;
-
-    /**
-     * Remove an attribute from each element in the set of matched elements.
-     *
-     * @param attributeName An attribute to remove; as of version 1.7, it can be a space-separated list of attributes.
-     */
-    removeAttr(attributeName: string): JQuery;
-
-    /**
-     * Remove a single class, multiple classes, or all classes from each element in the set of matched elements.
-     *
-     * @param className One or more space-separated classes to be removed from the class attribute of each matched element.
-     */
-    removeClass(className?: string): JQuery;
-    /**
-     * Remove a single class, multiple classes, or all classes from each element in the set of matched elements.
-     *
-     * @param function A function returning one or more space-separated class names to be removed. Receives the index position of the element in the set and the old class value as arguments.
-     */
-    removeClass(func: (index: number, className: string) => string): JQuery;
-
-    /**
-     * Remove a property for the set of matched elements.
-     *
-     * @param propertyName The name of the property to remove.
-     */
-    removeProp(propertyName: string): JQuery;
-
-    /**
-     * Add or remove one or more classes from each element in the set of matched elements, depending on either the class's presence or the value of the switch argument.
-     *
-     * @param className One or more class names (separated by spaces) to be toggled for each element in the matched set.
-     * @param swtch A Boolean (not just truthy/falsy) value to determine whether the class should be added or removed.
-     */
-    toggleClass(className: string, swtch?: boolean): JQuery;
-    /**
-     * Add or remove one or more classes from each element in the set of matched elements, depending on either the class's presence or the value of the switch argument.
-     *
-     * @param swtch A boolean value to determine whether the class should be added or removed.
-     */
-    toggleClass(swtch?: boolean): JQuery;
-    /**
-     * Add or remove one or more classes from each element in the set of matched elements, depending on either the class's presence or the value of the switch argument.
-     *
-     * @param func A function that returns class names to be toggled in the class attribute of each element in the matched set. Receives the index position of the element in the set, the old class value, and the switch as arguments.
-     * @param swtch A boolean value to determine whether the class should be added or removed.
-     */
-    toggleClass(func: (index: number, className: string, swtch: boolean) => string, swtch?: boolean): JQuery;
-
-    /**
-     * Get the current value of the first element in the set of matched elements.
-     */
-    val(): any;
-    /**
-     * Set the value of each element in the set of matched elements.
-     *
-     * @param value A string of text, an array of strings or number corresponding to the value of each matched element to set as selected/checked.
-     */
-    val(value: string|string[]|number): JQuery;
-    /**
-     * Set the value of each element in the set of matched elements.
-     *
-     * @param func A function returning the value to set. this is the current element. Receives the index position of the element in the set and the old value as arguments.
-     */
-    val(func: (index: number, value: string) => string): JQuery;
-
-
-    /**
-     * Get the value of style properties for the first element in the set of matched elements.
-     *
-     * @param propertyName A CSS property.
-     */
-    css(propertyName: string): string;
-    /**
-     * Set one or more CSS properties for the set of matched elements.
-     *
-     * @param propertyName A CSS property name.
-     * @param value A value to set for the property.
-     */
-    css(propertyName: string, value: string|number): JQuery;
-    /**
-     * Set one or more CSS properties for the set of matched elements.
-     *
-     * @param propertyName A CSS property name.
-     * @param value A function returning the value to set. this is the current element. Receives the index position of the element in the set and the old value as arguments.
-     */
-    css(propertyName: string, value: (index: number, value: string) => string|number): JQuery;
-    /**
-     * Set one or more CSS properties for the set of matched elements.
-     *
-     * @param properties An object of property-value pairs to set.
-     */
-    css(properties: Object): JQuery;
-
-    /**
-     * Get the current computed height for the first element in the set of matched elements.
-     */
-    height(): number;
-    /**
-     * Set the CSS height of every matched element.
-     *
-     * @param value An integer representing the number of pixels, or an integer with an optional unit of measure appended (as a string).
-     */
-    height(value: number|string): JQuery;
-    /**
-     * Set the CSS height of every matched element.
-     *
-     * @param func A function returning the height to set. Receives the index position of the element in the set and the old height as arguments. Within the function, this refers to the current element in the set.
-     */
-    height(func: (index: number, height: number) => number|string): JQuery;
-
-    /**
-     * Get the current computed height for the first element in the set of matched elements, including padding but not border.
-     */
-    innerHeight(): number;
-
-    /**
-     * Sets the inner height on elements in the set of matched elements, including padding but not border.
-     *
-     * @param value An integer representing the number of pixels, or an integer along with an optional unit of measure appended (as a string).
-     */
-    innerHeight(height: number|string): JQuery;
-    
-    /**
-     * Get the current computed width for the first element in the set of matched elements, including padding but not border.
-     */
-    innerWidth(): number;
-
-    /**
-     * Sets the inner width on elements in the set of matched elements, including padding but not border.
-     *
-     * @param value An integer representing the number of pixels, or an integer along with an optional unit of measure appended (as a string).
-     */
-    innerWidth(width: number|string): JQuery;
-    
-    /**
-     * Get the current coordinates of the first element in the set of matched elements, relative to the document.
-     */
-    offset(): JQueryCoordinates;
-    /**
-     * An object containing the properties top and left, which are integers indicating the new top and left coordinates for the elements.
-     *
-     * @param coordinates An object containing the properties top and left, which are integers indicating the new top and left coordinates for the elements.
-     */
-    offset(coordinates: JQueryCoordinates): JQuery;
-    /**
-     * An object containing the properties top and left, which are integers indicating the new top and left coordinates for the elements.
-     *
-     * @param func A function to return the coordinates to set. Receives the index of the element in the collection as the first argument and the current coordinates as the second argument. The function should return an object with the new top and left properties.
-     */
-    offset(func: (index: number, coords: JQueryCoordinates) => JQueryCoordinates): JQuery;
-
-    /**
-     * Get the current computed height for the first element in the set of matched elements, including padding, border, and optionally margin. Returns an integer (without "px") representation of the value or null if called on an empty set of elements.
-     *
-     * @param includeMargin A Boolean indicating whether to include the element's margin in the calculation.
-     */
-    outerHeight(includeMargin?: boolean): number;
-
-    /**
-     * Sets the outer height on elements in the set of matched elements, including padding and border.
-     *
-     * @param value An integer representing the number of pixels, or an integer along with an optional unit of measure appended (as a string).
-     */
-    outerHeight(height: number|string): JQuery;
-
-    /**
-     * Get the current computed width for the first element in the set of matched elements, including padding and border.
-     *
-     * @param includeMargin A Boolean indicating whether to include the element's margin in the calculation.
-     */
-    outerWidth(includeMargin?: boolean): number;
-
-    /**
-     * Sets the outer width on elements in the set of matched elements, including padding and border.
-     *
-     * @param value An integer representing the number of pixels, or an integer along with an optional unit of measure appended (as a string).
-     */
-    outerWidth(width: number|string): JQuery;
-
-    /**
-     * Get the current coordinates of the first element in the set of matched elements, relative to the offset parent.
-     */
-    position(): JQueryCoordinates;
-
-    /**
-     * Get the current horizontal position of the scroll bar for the first element in the set of matched elements or set the horizontal position of the scroll bar for every matched element.
-     */
-    scrollLeft(): number;
-    /**
-     * Set the current horizontal position of the scroll bar for each of the set of matched elements.
-     *
-     * @param value An integer indicating the new position to set the scroll bar to.
-     */
-    scrollLeft(value: number): JQuery;
-
-    /**
-     * Get the current vertical position of the scroll bar for the first element in the set of matched elements or set the vertical position of the scroll bar for every matched element.
-     */
-    scrollTop(): number;
-    /**
-     * Set the current vertical position of the scroll bar for each of the set of matched elements.
-     *
-     * @param value An integer indicating the new position to set the scroll bar to.
-     */
-    scrollTop(value: number): JQuery;
-
-    /**
-     * Get the current computed width for the first element in the set of matched elements.
-     */
-    width(): number;
-    /**
-     * Set the CSS width of each element in the set of matched elements.
-     *
-     * @param value An integer representing the number of pixels, or an integer along with an optional unit of measure appended (as a string).
-     */
-    width(value: number|string): JQuery;
-    /**
-     * Set the CSS width of each element in the set of matched elements.
-     *
-     * @param func A function returning the width to set. Receives the index position of the element in the set and the old width as arguments. Within the function, this refers to the current element in the set.
-     */
-    width(func: (index: number, width: number) => number|string): JQuery;
-
-    /**
-     * Remove from the queue all items that have not yet been run.
-     *
-     * @param queueName A string containing the name of the queue. Defaults to fx, the standard effects queue.
-     */
-    clearQueue(queueName?: string): JQuery;
-
-    /**
-     * Store arbitrary data associated with the matched elements.
-     *
-     * @param key A string naming the piece of data to set.
-     * @param value The new data value; it can be any Javascript type including Array or Object.
-     */
-    data(key: string, value: any): JQuery;
-    /**
-     * Return the value at the named data store for the first element in the jQuery collection, as set by data(name, value) or by an HTML5 data-* attribute.
-     *
-     * @param key Name of the data stored.
-     */
-    data(key: string): any;
-    /**
-     * Store arbitrary data associated with the matched elements.
-     *
-     * @param obj An object of key-value pairs of data to update.
-     */
-    data(obj: { [key: string]: any; }): JQuery;
-    /**
-     * Return the value at the named data store for the first element in the jQuery collection, as set by data(name, value) or by an HTML5 data-* attribute.
-     */
-    data(): any;
-
-    /**
-     * Execute the next function on the queue for the matched elements.
-     *
-     * @param queueName A string containing the name of the queue. Defaults to fx, the standard effects queue.
-     */
-    dequeue(queueName?: string): JQuery;
-
-    /**
-     * Remove a previously-stored piece of data.
-     *
-     * @param name A string naming the piece of data to delete or space-separated string naming the pieces of data to delete.
-     */
-    removeData(name: string): JQuery;
-    /**
-     * Remove a previously-stored piece of data.
-     *
-     * @param list An array of strings naming the pieces of data to delete.
-     */
-    removeData(list: string[]): JQuery;
-    /**
-     * Remove all previously-stored piece of data.
-     */
-    removeData(): JQuery;
-
-    /**
-     * Return a Promise object to observe when all actions of a certain type bound to the collection, queued or not, have finished.
-     *
-     * @param type The type of queue that needs to be observed. (default: fx)
-     * @param target Object onto which the promise methods have to be attached
-     */
-    promise(type?: string, target?: Object): JQueryPromise<any>;
-
-    /**
-     * Perform a custom animation of a set of CSS properties.
-     *
-     * @param properties An object of CSS properties and values that the animation will move toward.
-     * @param duration A string or number determining how long the animation will run.
-     * @param complete A function to call once the animation is complete.
-     */
-    animate(properties: Object, duration?: string|number, complete?: Function): JQuery;
-    /**
-     * Perform a custom animation of a set of CSS properties.
-     *
-     * @param properties An object of CSS properties and values that the animation will move toward.
-     * @param duration A string or number determining how long the animation will run.
-     * @param easing A string indicating which easing function to use for the transition. (default: swing)
-     * @param complete A function to call once the animation is complete.
-     */
-    animate(properties: Object, duration?: string|number, easing?: string, complete?: Function): JQuery;
-    /**
-     * Perform a custom animation of a set of CSS properties.
-     *
-     * @param properties An object of CSS properties and values that the animation will move toward.
-     * @param options A map of additional options to pass to the method.
-     */
-    animate(properties: Object, options: JQueryAnimationOptions): JQuery;
-
-    /**
-     * Set a timer to delay execution of subsequent items in the queue.
-     *
-     * @param duration An integer indicating the number of milliseconds to delay execution of the next item in the queue.
-     * @param queueName A string containing the name of the queue. Defaults to fx, the standard effects queue.
-     */
-    delay(duration: number, queueName?: string): JQuery;
-
-    /**
-     * Display the matched elements by fading them to opaque.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param complete A function to call once the animation is complete.
-     */
-    fadeIn(duration?: number|string, complete?: Function): JQuery;
-    /**
-     * Display the matched elements by fading them to opaque.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param easing A string indicating which easing function to use for the transition.
-     * @param complete A function to call once the animation is complete.
-     */
-    fadeIn(duration?: number|string, easing?: string, complete?: Function): JQuery;
-    /**
-     * Display the matched elements by fading them to opaque.
-     *
-     * @param options A map of additional options to pass to the method.
-     */
-    fadeIn(options: JQueryAnimationOptions): JQuery;
-
-    /**
-     * Hide the matched elements by fading them to transparent.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param complete A function to call once the animation is complete.
-     */
-    fadeOut(duration?: number|string, complete?: Function): JQuery;
-    /**
-     * Hide the matched elements by fading them to transparent.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param easing A string indicating which easing function to use for the transition.
-     * @param complete A function to call once the animation is complete.
-     */
-    fadeOut(duration?: number|string, easing?: string, complete?: Function): JQuery;
-    /**
-     * Hide the matched elements by fading them to transparent.
-     *
-     * @param options A map of additional options to pass to the method.
-     */
-    fadeOut(options: JQueryAnimationOptions): JQuery;
-
-    /**
-     * Adjust the opacity of the matched elements.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param opacity A number between 0 and 1 denoting the target opacity.
-     * @param complete A function to call once the animation is complete.
-     */
-    fadeTo(duration: string|number, opacity: number, complete?: Function): JQuery;
-    /**
-     * Adjust the opacity of the matched elements.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param opacity A number between 0 and 1 denoting the target opacity.
-     * @param easing A string indicating which easing function to use for the transition.
-     * @param complete A function to call once the animation is complete.
-     */
-    fadeTo(duration: string|number, opacity: number, easing?: string, complete?: Function): JQuery;
-
-    /**
-     * Display or hide the matched elements by animating their opacity.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param complete A function to call once the animation is complete.
-     */
-    fadeToggle(duration?: number|string, complete?: Function): JQuery;
-    /**
-     * Display or hide the matched elements by animating their opacity.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param easing A string indicating which easing function to use for the transition.
-     * @param complete A function to call once the animation is complete.
-     */
-    fadeToggle(duration?: number|string, easing?: string, complete?: Function): JQuery;
-    /**
-     * Display or hide the matched elements by animating their opacity.
-     *
-     * @param options A map of additional options to pass to the method.
-     */
-    fadeToggle(options: JQueryAnimationOptions): JQuery;
-
-    /**
-     * Stop the currently-running animation, remove all queued animations, and complete all animations for the matched elements.
-     *
-     * @param queue The name of the queue in which to stop animations.
-     */
-    finish(queue?: string): JQuery;
-
-    /**
-     * Hide the matched elements.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param complete A function to call once the animation is complete.
-     */
-    hide(duration?: number|string, complete?: Function): JQuery;
-    /**
-     * Hide the matched elements.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param easing A string indicating which easing function to use for the transition.
-     * @param complete A function to call once the animation is complete.
-     */
-    hide(duration?: number|string, easing?: string, complete?: Function): JQuery;
-    /**
-     * Hide the matched elements.
-     *
-     * @param options A map of additional options to pass to the method.
-     */
-    hide(options: JQueryAnimationOptions): JQuery;
-
-    /**
-     * Display the matched elements.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param complete A function to call once the animation is complete.
-     */
-    show(duration?: number|string, complete?: Function): JQuery;
-    /**
-     * Display the matched elements.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param easing A string indicating which easing function to use for the transition.
-     * @param complete A function to call once the animation is complete.
-     */
-    show(duration?: number|string, easing?: string, complete?: Function): JQuery;
-    /**
-     * Display the matched elements.
-     *
-     * @param options A map of additional options to pass to the method.
-     */
-    show(options: JQueryAnimationOptions): JQuery;
-
-    /**
-     * Display the matched elements with a sliding motion.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param complete A function to call once the animation is complete.
-     */
-    slideDown(duration?: number|string, complete?: Function): JQuery;
-    /**
-     * Display the matched elements with a sliding motion.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param easing A string indicating which easing function to use for the transition.
-     * @param complete A function to call once the animation is complete.
-     */
-    slideDown(duration?: number|string, easing?: string, complete?: Function): JQuery;
-    /**
-     * Display the matched elements with a sliding motion.
-     *
-     * @param options A map of additional options to pass to the method.
-     */
-    slideDown(options: JQueryAnimationOptions): JQuery;
-
-    /**
-     * Display or hide the matched elements with a sliding motion.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param complete A function to call once the animation is complete.
-     */
-    slideToggle(duration?: number|string, complete?: Function): JQuery;
-    /**
-     * Display or hide the matched elements with a sliding motion.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param easing A string indicating which easing function to use for the transition.
-     * @param complete A function to call once the animation is complete.
-     */
-    slideToggle(duration?: number|string, easing?: string, complete?: Function): JQuery;
-    /**
-     * Display or hide the matched elements with a sliding motion.
-     *
-     * @param options A map of additional options to pass to the method.
-     */
-    slideToggle(options: JQueryAnimationOptions): JQuery;
-
-    /**
-     * Hide the matched elements with a sliding motion.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param complete A function to call once the animation is complete.
-     */
-    slideUp(duration?: number|string, complete?: Function): JQuery;
-    /**
-     * Hide the matched elements with a sliding motion.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param easing A string indicating which easing function to use for the transition.
-     * @param complete A function to call once the animation is complete.
-     */
-    slideUp(duration?: number|string, easing?: string, complete?: Function): JQuery;
-    /**
-     * Hide the matched elements with a sliding motion.
-     *
-     * @param options A map of additional options to pass to the method.
-     */
-    slideUp(options: JQueryAnimationOptions): JQuery;
-
-    /**
-     * Stop the currently-running animation on the matched elements.
-     *
-     * @param clearQueue A Boolean indicating whether to remove queued animation as well. Defaults to false.
-     * @param jumpToEnd A Boolean indicating whether to complete the current animation immediately. Defaults to false.
-     */
-    stop(clearQueue?: boolean, jumpToEnd?: boolean): JQuery;
-    /**
-     * Stop the currently-running animation on the matched elements.
-     *
-     * @param queue The name of the queue in which to stop animations.
-     * @param clearQueue A Boolean indicating whether to remove queued animation as well. Defaults to false.
-     * @param jumpToEnd A Boolean indicating whether to complete the current animation immediately. Defaults to false.
-     */
-    stop(queue?: string, clearQueue?: boolean, jumpToEnd?: boolean): JQuery;
-
-    /**
-     * Display or hide the matched elements.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param complete A function to call once the animation is complete.
-     */
-    toggle(duration?: number|string, complete?: Function): JQuery;
-    /**
-     * Display or hide the matched elements.
-     *
-     * @param duration A string or number determining how long the animation will run.
-     * @param easing A string indicating which easing function to use for the transition.
-     * @param complete A function to call once the animation is complete.
-     */
-    toggle(duration?: number|string, easing?: string, complete?: Function): JQuery;
-    /**
-     * Display or hide the matched elements.
-     *
-     * @param options A map of additional options to pass to the method.
-     */
-    toggle(options: JQueryAnimationOptions): JQuery;
-    /**
-     * Display or hide the matched elements.
-     *
-     * @param showOrHide A Boolean indicating whether to show or hide the elements.
-     */
-    toggle(showOrHide: boolean): JQuery;
-
-    /**
-     * Attach a handler to an event for the elements.
-     * 
-     * @param eventType A string containing one or more DOM event types, such as "click" or "submit," or custom event names.
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    bind(eventType: string, eventData: any, handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Attach a handler to an event for the elements.
-     * 
-     * @param eventType A string containing one or more DOM event types, such as "click" or "submit," or custom event names.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    bind(eventType: string, handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Attach a handler to an event for the elements.
-     * 
-     * @param eventType A string containing one or more DOM event types, such as "click" or "submit," or custom event names.
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param preventBubble Setting the third argument to false will attach a function that prevents the default action from occurring and stops the event from bubbling. The default is true.
-     */
-    bind(eventType: string, eventData: any, preventBubble: boolean): JQuery;
-    /**
-     * Attach a handler to an event for the elements.
-     * 
-     * @param eventType A string containing one or more DOM event types, such as "click" or "submit," or custom event names.
-     * @param preventBubble Setting the third argument to false will attach a function that prevents the default action from occurring and stops the event from bubbling. The default is true.
-     */
-    bind(eventType: string, preventBubble: boolean): JQuery;
-    /**
-     * Attach a handler to an event for the elements.
-     * 
-     * @param events An object containing one or more DOM event types and functions to execute for them.
-     */
-    bind(events: any): JQuery;
-
-    /**
-     * Trigger the "blur" event on an element
-     */
-    blur(): JQuery;
-    /**
-     * Bind an event handler to the "blur" JavaScript event
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    blur(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "blur" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    blur(eventData?: any, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "change" event on an element.
-     */
-    change(): JQuery;
-    /**
-     * Bind an event handler to the "change" JavaScript event
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    change(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "change" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    change(eventData?: any, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "click" event on an element.
-     */
-    click(): JQuery;
-    /**
-     * Bind an event handler to the "click" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     */
-    click(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "click" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    click(eventData?: any, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "dblclick" event on an element.
-     */
-    dblclick(): JQuery;
-    /**
-     * Bind an event handler to the "dblclick" JavaScript event
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    dblclick(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "dblclick" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    dblclick(eventData?: any, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-
-    delegate(selector: any, eventType: string, handler: (eventObject: JQueryEventObject) => any): JQuery;
-    delegate(selector: any, eventType: string, eventData: any, handler: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "focus" event on an element.
-     */
-    focus(): JQuery;
-    /**
-     * Bind an event handler to the "focus" JavaScript event
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    focus(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "focus" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    focus(eventData?: any, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "focusin" event on an element.
-     */
-    focusin(): JQuery;
-    /**
-     * Bind an event handler to the "focusin" JavaScript event
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    focusin(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "focusin" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    focusin(eventData: Object, handler: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "focusout" event on an element.
-     */
-    focusout(): JQuery;
-    /**
-     * Bind an event handler to the "focusout" JavaScript event
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    focusout(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "focusout" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    focusout(eventData: Object, handler: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Bind two handlers to the matched elements, to be executed when the mouse pointer enters and leaves the elements.
-     *
-     * @param handlerIn A function to execute when the mouse pointer enters the element.
-     * @param handlerOut A function to execute when the mouse pointer leaves the element.
-     */
-    hover(handlerIn: (eventObject: JQueryEventObject) => any, handlerOut: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind a single handler to the matched elements, to be executed when the mouse pointer enters or leaves the elements.
-     *
-     * @param handlerInOut A function to execute when the mouse pointer enters or leaves the element.
-     */
-    hover(handlerInOut: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "keydown" event on an element.
-     */
-    keydown(): JQuery;
-    /**
-     * Bind an event handler to the "keydown" JavaScript event
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    keydown(handler: (eventObject: JQueryKeyEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "keydown" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    keydown(eventData?: any, handler?: (eventObject: JQueryKeyEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "keypress" event on an element.
-     */
-    keypress(): JQuery;
-    /**
-     * Bind an event handler to the "keypress" JavaScript event
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    keypress(handler: (eventObject: JQueryKeyEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "keypress" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    keypress(eventData?: any, handler?: (eventObject: JQueryKeyEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "keyup" event on an element.
-     */
-    keyup(): JQuery;
-    /**
-     * Bind an event handler to the "keyup" JavaScript event
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    keyup(handler: (eventObject: JQueryKeyEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "keyup" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    keyup(eventData?: any, handler?: (eventObject: JQueryKeyEventObject) => any): JQuery;
-
-    /**
-     * Bind an event handler to the "load" JavaScript event.
-     *
-     * @param handler A function to execute when the event is triggered.
-     */
-    load(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "load" JavaScript event.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute when the event is triggered.
-     */
-    load(eventData?: any, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "mousedown" event on an element.
-     */
-    mousedown(): JQuery;
-    /**
-     * Bind an event handler to the "mousedown" JavaScript event.
-     *
-     * @param handler A function to execute when the event is triggered.
-     */
-    mousedown(handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "mousedown" JavaScript event.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute when the event is triggered.
-     */
-    mousedown(eventData: Object, handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "mouseenter" event on an element.
-     */
-    mouseenter(): JQuery;
-    /**
-     * Bind an event handler to be fired when the mouse enters an element.
-     *
-     * @param handler A function to execute when the event is triggered.
-     */
-    mouseenter(handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to be fired when the mouse enters an element.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute when the event is triggered.
-     */
-    mouseenter(eventData: Object, handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "mouseleave" event on an element.
-     */
-    mouseleave(): JQuery;
-    /**
-     * Bind an event handler to be fired when the mouse leaves an element.
-     *
-     * @param handler A function to execute when the event is triggered.
-     */
-    mouseleave(handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to be fired when the mouse leaves an element.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute when the event is triggered.
-     */
-    mouseleave(eventData: Object, handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "mousemove" event on an element.
-     */
-    mousemove(): JQuery;
-    /**
-     * Bind an event handler to the "mousemove" JavaScript event.
-     *
-     * @param handler A function to execute when the event is triggered.
-     */
-    mousemove(handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "mousemove" JavaScript event.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute when the event is triggered.
-     */
-    mousemove(eventData: Object, handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "mouseout" event on an element.
-     */
-    mouseout(): JQuery;
-    /**
-     * Bind an event handler to the "mouseout" JavaScript event.
-     *
-     * @param handler A function to execute when the event is triggered.
-     */
-    mouseout(handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "mouseout" JavaScript event.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute when the event is triggered.
-     */
-    mouseout(eventData: Object, handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "mouseover" event on an element.
-     */
-    mouseover(): JQuery;
-    /**
-     * Bind an event handler to the "mouseover" JavaScript event.
-     *
-     * @param handler A function to execute when the event is triggered.
-     */
-    mouseover(handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "mouseover" JavaScript event.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute when the event is triggered.
-     */
-    mouseover(eventData: Object, handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "mouseup" event on an element.
-     */
-    mouseup(): JQuery;
-    /**
-     * Bind an event handler to the "mouseup" JavaScript event.
-     *
-     * @param handler A function to execute when the event is triggered.
-     */
-    mouseup(handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "mouseup" JavaScript event.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute when the event is triggered.
-     */
-    mouseup(eventData: Object, handler: (eventObject: JQueryMouseEventObject) => any): JQuery;
-
-    /**
-     * Remove an event handler.
-     */
-    off(): JQuery;
-    /**
-     * Remove an event handler.
-     *
-     * @param events One or more space-separated event types and optional namespaces, or just namespaces, such as "click", "keydown.myPlugin", or ".myPlugin".
-     * @param selector A selector which should match the one originally passed to .on() when attaching event handlers.
-     * @param handler A handler function previously attached for the event(s), or the special value false.
-     */
-    off(events: string, selector?: string, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Remove an event handler.
-     *
-     * @param events One or more space-separated event types and optional namespaces, or just namespaces, such as "click", "keydown.myPlugin", or ".myPlugin".
-     * @param handler A handler function previously attached for the event(s), or the special value false. Takes handler with extra args that can be attached with on().
-     */
-    off(events: string, handler: (eventObject: JQueryEventObject, ...args: any[]) => any): JQuery;
-    /**
-     * Remove an event handler.
-     *
-     * @param events One or more space-separated event types and optional namespaces, or just namespaces, such as "click", "keydown.myPlugin", or ".myPlugin".
-     * @param handler A handler function previously attached for the event(s), or the special value false.
-     */
-    off(events: string, handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Remove an event handler.
-     *
-     * @param events An object where the string keys represent one or more space-separated event types and optional namespaces, and the values represent handler functions previously attached for the event(s).
-     * @param selector A selector which should match the one originally passed to .on() when attaching event handlers.
-     */
-    off(events: { [key: string]: any; }, selector?: string): JQuery;
-
-    /**
-     * Attach an event handler function for one or more events to the selected elements.
-     *
-     * @param events One or more space-separated event types and optional namespaces, such as "click" or "keydown.myPlugin".
-     * @param handler A function to execute when the event is triggered. The value false is also allowed as a shorthand for a function that simply does return false. Rest parameter args is for optional parameters passed to jQuery.trigger(). Note that the actual parameters on the event handler function must be marked as optional (? syntax).
-     */
-    on(events: string, handler: (eventObject: JQueryEventObject, ...args: any[]) => any): JQuery;
-    /**
-     * Attach an event handler function for one or more events to the selected elements.
-     *
-     * @param events One or more space-separated event types and optional namespaces, such as "click" or "keydown.myPlugin".
-     * @param data Data to be passed to the handler in event.data when an event is triggered.
-     * @param handler A function to execute when the event is triggered. The value false is also allowed as a shorthand for a function that simply does return false.
-    */
-    on(events: string, data : any, handler: (eventObject: JQueryEventObject, ...args: any[]) => any): JQuery;
-    /**
-     * Attach an event handler function for one or more events to the selected elements.
-     *
-     * @param events One or more space-separated event types and optional namespaces, such as "click" or "keydown.myPlugin".
-     * @param selector A selector string to filter the descendants of the selected elements that trigger the event. If the selector is null or omitted, the event is always triggered when it reaches the selected element.
-     * @param handler A function to execute when the event is triggered. The value false is also allowed as a shorthand for a function that simply does return false.
-     */
-    on(events: string, selector: string, handler: (eventObject: JQueryEventObject, ...eventData: any[]) => any): JQuery;
-    /**
-     * Attach an event handler function for one or more events to the selected elements.
-     *
-     * @param events One or more space-separated event types and optional namespaces, such as "click" or "keydown.myPlugin".
-     * @param selector A selector string to filter the descendants of the selected elements that trigger the event. If the selector is null or omitted, the event is always triggered when it reaches the selected element.
-     * @param data Data to be passed to the handler in event.data when an event is triggered.
-     * @param handler A function to execute when the event is triggered. The value false is also allowed as a shorthand for a function that simply does return false.
-     */
-    on(events: string, selector: string, data: any, handler: (eventObject: JQueryEventObject, ...eventData: any[]) => any): JQuery;
-    /**
-     * Attach an event handler function for one or more events to the selected elements.
-     *
-     * @param events An object in which the string keys represent one or more space-separated event types and optional namespaces, and the values represent a handler function to be called for the event(s).
-     * @param selector A selector string to filter the descendants of the selected elements that will call the handler. If the selector is null or omitted, the handler is always called when it reaches the selected element.
-     * @param data Data to be passed to the handler in event.data when an event occurs.
-     */
-    on(events: { [key: string]: any; }, selector?: string, data?: any): JQuery;
-    /**
-     * Attach an event handler function for one or more events to the selected elements.
-     *
-     * @param events An object in which the string keys represent one or more space-separated event types and optional namespaces, and the values represent a handler function to be called for the event(s).
-     * @param data Data to be passed to the handler in event.data when an event occurs.
-     */
-    on(events: { [key: string]: any; }, data?: any): JQuery;
-
-    /**
-     * Attach a handler to an event for the elements. The handler is executed at most once per element per event type.
-     *
-     * @param events A string containing one or more JavaScript event types, such as "click" or "submit," or custom event names.
-     * @param handler A function to execute at the time the event is triggered.
-     */
-    one(events: string, handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Attach a handler to an event for the elements. The handler is executed at most once per element per event type.
-     *
-     * @param events A string containing one or more JavaScript event types, such as "click" or "submit," or custom event names.
-     * @param data An object containing data that will be passed to the event handler.
-     * @param handler A function to execute at the time the event is triggered.
-     */
-    one(events: string, data: Object, handler: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Attach a handler to an event for the elements. The handler is executed at most once per element per event type.
-     *
-     * @param events One or more space-separated event types and optional namespaces, such as "click" or "keydown.myPlugin".
-     * @param selector A selector string to filter the descendants of the selected elements that trigger the event. If the selector is null or omitted, the event is always triggered when it reaches the selected element.
-     * @param handler A function to execute when the event is triggered. The value false is also allowed as a shorthand for a function that simply does return false.
-     */
-    one(events: string, selector: string, handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Attach a handler to an event for the elements. The handler is executed at most once per element per event type.
-     *
-     * @param events One or more space-separated event types and optional namespaces, such as "click" or "keydown.myPlugin".
-     * @param selector A selector string to filter the descendants of the selected elements that trigger the event. If the selector is null or omitted, the event is always triggered when it reaches the selected element.
-     * @param data Data to be passed to the handler in event.data when an event is triggered.
-     * @param handler A function to execute when the event is triggered. The value false is also allowed as a shorthand for a function that simply does return false.
-     */
-    one(events: string, selector: string, data: any, handler: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Attach a handler to an event for the elements. The handler is executed at most once per element per event type.
-     *
-     * @param events An object in which the string keys represent one or more space-separated event types and optional namespaces, and the values represent a handler function to be called for the event(s).
-     * @param selector A selector string to filter the descendants of the selected elements that will call the handler. If the selector is null or omitted, the handler is always called when it reaches the selected element.
-     * @param data Data to be passed to the handler in event.data when an event occurs.
-     */
-    one(events: { [key: string]: any; }, selector?: string, data?: any): JQuery;
-
-    /**
-     * Attach a handler to an event for the elements. The handler is executed at most once per element per event type.
-     *
-     * @param events An object in which the string keys represent one or more space-separated event types and optional namespaces, and the values represent a handler function to be called for the event(s).
-     * @param data Data to be passed to the handler in event.data when an event occurs.
-     */
-    one(events: { [key: string]: any; }, data?: any): JQuery;
-
-
-    /**
-     * Specify a function to execute when the DOM is fully loaded.
-     *
-     * @param handler A function to execute after the DOM is ready.
-     */
-    ready(handler: (jQueryAlias?: JQueryStatic) => any): JQuery;
-
-    /**
-     * Trigger the "resize" event on an element.
-     */
-    resize(): JQuery;
-    /**
-     * Bind an event handler to the "resize" JavaScript event.
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    resize(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "resize" JavaScript event.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    resize(eventData: Object, handler: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "scroll" event on an element.
-     */
-    scroll(): JQuery;
-    /**
-     * Bind an event handler to the "scroll" JavaScript event.
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    scroll(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "scroll" JavaScript event.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    scroll(eventData: Object, handler: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "select" event on an element.
-     */
-    select(): JQuery;
-    /**
-     * Bind an event handler to the "select" JavaScript event.
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    select(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "select" JavaScript event.
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    select(eventData: Object, handler: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Trigger the "submit" event on an element.
-     */
-    submit(): JQuery;
-    /**
-     * Bind an event handler to the "submit" JavaScript event
-     *
-     * @param handler A function to execute each time the event is triggered.
-     */
-    submit(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "submit" JavaScript event
-     *
-     * @param eventData An object containing data that will be passed to the event handler.
-     * @param handler A function to execute each time the event is triggered.
-     */
-    submit(eventData?: any, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Execute all handlers and behaviors attached to the matched elements for the given event type.
-     * 
-     * @param eventType A string containing a JavaScript event type, such as click or submit.
-     * @param extraParameters Additional parameters to pass along to the event handler.
-     */
-    trigger(eventType: string, extraParameters?: any[]|Object): JQuery;
-    /**
-     * Execute all handlers and behaviors attached to the matched elements for the given event type.
-     * 
-     * @param event A jQuery.Event object.
-     * @param extraParameters Additional parameters to pass along to the event handler.
-     */
-    trigger(event: JQueryEventObject, extraParameters?: any[]|Object): JQuery;
-
-    /**
-     * Execute all handlers attached to an element for an event.
-     * 
-     * @param eventType A string containing a JavaScript event type, such as click or submit.
-     * @param extraParameters An array of additional parameters to pass along to the event handler.
-     */
-    triggerHandler(eventType: string, ...extraParameters: any[]): Object;
-
-    /**
-     * Execute all handlers attached to an element for an event.
-     * 
-     * @param event A jQuery.Event object.
-     * @param extraParameters An array of additional parameters to pass along to the event handler.
-     */
-    triggerHandler(event: JQueryEventObject, ...extraParameters: any[]): Object;
-
-    /**
-     * Remove a previously-attached event handler from the elements.
-     * 
-     * @param eventType A string containing a JavaScript event type, such as click or submit.
-     * @param handler The function that is to be no longer executed.
-     */
-    unbind(eventType?: string, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Remove a previously-attached event handler from the elements.
-     * 
-     * @param eventType A string containing a JavaScript event type, such as click or submit.
-     * @param fls Unbinds the corresponding 'return false' function that was bound using .bind( eventType, false ).
-     */
-    unbind(eventType: string, fls: boolean): JQuery;
-    /**
-     * Remove a previously-attached event handler from the elements.
-     * 
-     * @param evt A JavaScript event object as passed to an event handler.
-     */
-    unbind(evt: any): JQuery;
-
-    /**
-     * Remove a handler from the event for all elements which match the current selector, based upon a specific set of root elements.
-     */
-    undelegate(): JQuery;
-    /**
-     * Remove a handler from the event for all elements which match the current selector, based upon a specific set of root elements.
-     * 
-     * @param selector A selector which will be used to filter the event results.
-     * @param eventType A string containing a JavaScript event type, such as "click" or "keydown"
-     * @param handler A function to execute at the time the event is triggered.
-     */
-    undelegate(selector: string, eventType: string, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Remove a handler from the event for all elements which match the current selector, based upon a specific set of root elements.
-     * 
-     * @param selector A selector which will be used to filter the event results.
-     * @param events An object of one or more event types and previously bound functions to unbind from them.
-     */
-    undelegate(selector: string, events: Object): JQuery;
-    /**
-     * Remove a handler from the event for all elements which match the current selector, based upon a specific set of root elements.
-     * 
-     * @param namespace A string containing a namespace to unbind all events from.
-     */
-    undelegate(namespace: string): JQuery;
-
-    /**
-     * Bind an event handler to the "unload" JavaScript event. (DEPRECATED from v1.8)
-     * 
-     * @param handler A function to execute when the event is triggered.
-     */
-    unload(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "unload" JavaScript event. (DEPRECATED from v1.8)
-     * 
-     * @param eventData A plain object of data that will be passed to the event handler.
-     * @param handler A function to execute when the event is triggered.
-     */
-    unload(eventData?: any, handler?: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * The DOM node context originally passed to jQuery(); if none was passed then context will likely be the document. (DEPRECATED from v1.10)
-     */
-    context: Element;
-
-    jquery: string;
-
-    /**
-     * Bind an event handler to the "error" JavaScript event. (DEPRECATED from v1.8)
-     * 
-     * @param handler A function to execute when the event is triggered.
-     */
-    error(handler: (eventObject: JQueryEventObject) => any): JQuery;
-    /**
-     * Bind an event handler to the "error" JavaScript event. (DEPRECATED from v1.8)
-     * 
-     * @param eventData A plain object of data that will be passed to the event handler.
-     * @param handler A function to execute when the event is triggered.
-     */
-    error(eventData: any, handler: (eventObject: JQueryEventObject) => any): JQuery;
-
-    /**
-     * Add a collection of DOM elements onto the jQuery stack.
-     * 
-     * @param elements An array of elements to push onto the stack and make into a new jQuery object.
-     */
-    pushStack(elements: any[]): JQuery;
-    /**
-     * Add a collection of DOM elements onto the jQuery stack.
-     * 
-     * @param elements An array of elements to push onto the stack and make into a new jQuery object.
-     * @param name The name of a jQuery method that generated the array of elements.
-     * @param arguments The arguments that were passed in to the jQuery method (for serialization).
-     */
-    pushStack(elements: any[], name: string, arguments: any[]): JQuery;
-
-    /**
-     * Insert content, specified by the parameter, after each element in the set of matched elements.
-     * 
-     * param content1 HTML string, DOM element, array of elements, or jQuery object to insert after each element in the set of matched elements.
-     * param content2 One or more additional DOM elements, arrays of elements, HTML strings, or jQuery objects to insert after each element in the set of matched elements.
-     */
-    after(content1: JQuery|any[]|Element|Text|string, ...content2: any[]): JQuery;
-    /**
-     * Insert content, specified by the parameter, after each element in the set of matched elements.
-     * 
-     * param func A function that returns an HTML string, DOM element(s), or jQuery object to insert after each element in the set of matched elements. Receives the index position of the element in the set as an argument. Within the function, this refers to the current element in the set.
-     */
-    after(func: (index: number, html: string) => string|Element|JQuery): JQuery;
-
-    /**
-     * Insert content, specified by the parameter, to the end of each element in the set of matched elements.
-     * 
-     * param content1 DOM element, array of elements, HTML string, or jQuery object to insert at the end of each element in the set of matched elements.
-     * param content2 One or more additional DOM elements, arrays of elements, HTML strings, or jQuery objects to insert at the end of each element in the set of matched elements.
-     */
-    append(content1: JQuery|any[]|Element|Text|string, ...content2: any[]): JQuery;
-    /**
-     * Insert content, specified by the parameter, to the end of each element in the set of matched elements.
-     * 
-     * param func A function that returns an HTML string, DOM element(s), or jQuery object to insert at the end of each element in the set of matched elements. Receives the index position of the element in the set and the old HTML value of the element as arguments. Within the function, this refers to the current element in the set.
-     */
-    append(func: (index: number, html: string) => string|Element|JQuery): JQuery;
-
-    /**
-     * Insert every element in the set of matched elements to the end of the target.
-     * 
-     * @param target A selector, element, HTML string, array of elements, or jQuery object; the matched set of elements will be inserted at the end of the element(s) specified by this parameter.
-     */
-    appendTo(target: JQuery|any[]|Element|string): JQuery;
-
-    /**
-     * Insert content, specified by the parameter, before each element in the set of matched elements.
-     * 
-     * param content1 HTML string, DOM element, array of elements, or jQuery object to insert before each element in the set of matched elements.
-     * param content2 One or more additional DOM elements, arrays of elements, HTML strings, or jQuery objects to insert before each element in the set of matched elements.
-     */
-    before(content1: JQuery|any[]|Element|Text|string, ...content2: any[]): JQuery;
-    /**
-     * Insert content, specified by the parameter, before each element in the set of matched elements.
-     * 
-     * param func A function that returns an HTML string, DOM element(s), or jQuery object to insert before each element in the set of matched elements. Receives the index position of the element in the set as an argument. Within the function, this refers to the current element in the set.
-     */
-    before(func: (index: number, html: string) => string|Element|JQuery): JQuery;
-
-    /**
-     * Create a deep copy of the set of matched elements.
-     * 
-     * param withDataAndEvents A Boolean indicating whether event handlers and data should be copied along with the elements. The default value is false.
-     * param deepWithDataAndEvents A Boolean indicating whether event handlers and data for all children of the cloned element should be copied. By default its value matches the first argument's value (which defaults to false).
-     */
-    clone(withDataAndEvents?: boolean, deepWithDataAndEvents?: boolean): JQuery;
-
-    /**
-     * Remove the set of matched elements from the DOM.
-     * 
-     * param selector A selector expression that filters the set of matched elements to be removed.
-     */
-    detach(selector?: string): JQuery;
-
-    /**
-     * Remove all child nodes of the set of matched elements from the DOM.
-     */
-    empty(): JQuery;
-
-    /**
-     * Insert every element in the set of matched elements after the target.
-     * 
-     * param target A selector, element, array of elements, HTML string, or jQuery object; the matched set of elements will be inserted after the element(s) specified by this parameter.
-     */
-    insertAfter(target: JQuery|any[]|Element|Text|string): JQuery;
-
-    /**
-     * Insert every element in the set of matched elements before the target.
-     * 
-     * param target A selector, element, array of elements, HTML string, or jQuery object; the matched set of elements will be inserted before the element(s) specified by this parameter.
-     */
-    insertBefore(target: JQuery|any[]|Element|Text|string): JQuery;
-
-    /**
-     * Insert content, specified by the parameter, to the beginning of each element in the set of matched elements.
-     * 
-     * param content1 DOM element, array of elements, HTML string, or jQuery object to insert at the beginning of each element in the set of matched elements.
-     * param content2 One or more additional DOM elements, arrays of elements, HTML strings, or jQuery objects to insert at the beginning of each element in the set of matched elements.
-     */
-    prepend(content1: JQuery|any[]|Element|Text|string, ...content2: any[]): JQuery;
-    /**
-     * Insert content, specified by the parameter, to the beginning of each element in the set of matched elements.
-     * 
-     * param func A function that returns an HTML string, DOM element(s), or jQuery object to insert at the beginning of each element in the set of matched elements. Receives the index position of the element in the set and the old HTML value of the element as arguments. Within the function, this refers to the current element in the set.
-     */
-    prepend(func: (index: number, html: string) => string|Element|JQuery): JQuery;
-
-    /**
-     * Insert every element in the set of matched elements to the beginning of the target.
-     * 
-     * @param target A selector, element, HTML string, array of elements, or jQuery object; the matched set of elements will be inserted at the beginning of the element(s) specified by this parameter.
-     */
-    prependTo(target: JQuery|any[]|Element|string): JQuery;
-
-    /**
-     * Remove the set of matched elements from the DOM.
-     * 
-     * @param selector A selector expression that filters the set of matched elements to be removed.
-     */
-    remove(selector?: string): JQuery;
-
-    /**
-     * Replace each target element with the set of matched elements.
-     * 
-     * @param target A selector string, jQuery object, DOM element, or array of elements indicating which element(s) to replace.
-     */
-    replaceAll(target: JQuery|any[]|Element|string): JQuery;
-
-    /**
-     * Replace each element in the set of matched elements with the provided new content and return the set of elements that was removed.
-     * 
-     * param newContent The content to insert. May be an HTML string, DOM element, array of DOM elements, or jQuery object.
-     */
-    replaceWith(newContent: JQuery|any[]|Element|Text|string): JQuery;
-    /**
-     * Replace each element in the set of matched elements with the provided new content and return the set of elements that was removed.
-     * 
-     * param func A function that returns content with which to replace the set of matched elements.
-     */
-    replaceWith(func: () => Element|JQuery): JQuery;
-
-    /**
-     * Get the combined text contents of each element in the set of matched elements, including their descendants.
-     */
-    text(): string;
-    /**
-     * Set the content of each element in the set of matched elements to the specified text.
-     * 
-     * @param text The text to set as the content of each matched element. When Number or Boolean is supplied, it will be converted to a String representation.
-     */
-    text(text: string|number|boolean): JQuery;
-    /**
-     * Set the content of each element in the set of matched elements to the specified text.
-     * 
-     * @param func A function returning the text content to set. Receives the index position of the element in the set and the old text value as arguments.
-     */
-    text(func: (index: number, text: string) => string): JQuery;
-
-    /**
-     * Retrieve all the elements contained in the jQuery set, as an array.
-     */
-    toArray(): any[];
-
-    /**
-     * Remove the parents of the set of matched elements from the DOM, leaving the matched elements in their place.
-     */
-    unwrap(): JQuery;
-
-    /**
-     * Wrap an HTML structure around each element in the set of matched elements.
-     * 
-     * @param wrappingElement A selector, element, HTML string, or jQuery object specifying the structure to wrap around the matched elements.
-     */
-    wrap(wrappingElement: JQuery|Element|string): JQuery;
-    /**
-     * Wrap an HTML structure around each element in the set of matched elements.
-     * 
-     * @param func A callback function returning the HTML content or jQuery object to wrap around the matched elements. Receives the index position of the element in the set as an argument. Within the function, this refers to the current element in the set.
-     */
-    wrap(func: (index: number) => string|JQuery): JQuery;
-
-    /**
-     * Wrap an HTML structure around all elements in the set of matched elements.
-     * 
-     * @param wrappingElement A selector, element, HTML string, or jQuery object specifying the structure to wrap around the matched elements.
-     */
-    wrapAll(wrappingElement: JQuery|Element|string): JQuery;
-    wrapAll(func: (index: number) => string): JQuery;
-
-    /**
-     * Wrap an HTML structure around the content of each element in the set of matched elements.
-     * 
-     * @param wrappingElement An HTML snippet, selector expression, jQuery object, or DOM element specifying the structure to wrap around the content of the matched elements.
-     */
-    wrapInner(wrappingElement: JQuery|Element|string): JQuery;
-    /**
-     * Wrap an HTML structure around the content of each element in the set of matched elements.
-     * 
-     * @param func A callback function which generates a structure to wrap around the content of the matched elements. Receives the index position of the element in the set as an argument. Within the function, this refers to the current element in the set.
-     */
-    wrapInner(func: (index: number) => string): JQuery;
-
-    /**
-     * Iterate over a jQuery object, executing a function for each matched element.
-     * 
-     * @param func A function to execute for each matched element.
-     */
-    each(func: (index: number, elem: Element) => any): JQuery;
-
-    /**
-     * Retrieve one of the elements matched by the jQuery object.
-     * 
-     * @param index A zero-based integer indicating which element to retrieve.
-     */
-    get(index: number): HTMLElement;
-    /**
-     * Retrieve the elements matched by the jQuery object.
-     */
-    get(): any[];
-
-    /**
-     * Search for a given element from among the matched elements.
-     */
-    index(): number;
-    /**
-     * Search for a given element from among the matched elements.
-     * 
-     * @param selector A selector representing a jQuery collection in which to look for an element.
-     */
-    index(selector: string|JQuery|Element): number;
-
-    /**
-     * The number of elements in the jQuery object.
-     */
-    length: number;
-    /**
-     * A selector representing selector passed to jQuery(), if any, when creating the original set.
-     * version deprecated: 1.7, removed: 1.9
-     */
-    selector: string;
-    [index: string]: any;
-    [index: number]: HTMLElement;
-
-    /**
-     * Add elements to the set of matched elements.
-     * 
-     * @param selector A string representing a selector expression to find additional elements to add to the set of matched elements.
-     * @param context The point in the document at which the selector should begin matching; similar to the context argument of the $(selector, context) method.
-     */
-    add(selector: string, context?: Element): JQuery;
-    /**
-     * Add elements to the set of matched elements.
-     * 
-     * @param elements One or more elements to add to the set of matched elements.
-     */
-    add(...elements: Element[]): JQuery;
-    /**
-     * Add elements to the set of matched elements.
-     * 
-     * @param html An HTML fragment to add to the set of matched elements.
-     */
-    add(html: string): JQuery;
-    /**
-     * Add elements to the set of matched elements.
-     * 
-     * @param obj An existing jQuery object to add to the set of matched elements.
-     */
-    add(obj: JQuery): JQuery;
-
-    /**
-     * Get the children of each element in the set of matched elements, optionally filtered by a selector.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    children(selector?: string): JQuery;
-
-    /**
-     * For each element in the set, get the first element that matches the selector by testing the element itself and traversing up through its ancestors in the DOM tree.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    closest(selector: string): JQuery;
-    /**
-     * For each element in the set, get the first element that matches the selector by testing the element itself and traversing up through its ancestors in the DOM tree.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     * @param context A DOM element within which a matching element may be found. If no context is passed in then the context of the jQuery set will be used instead.
-     */
-    closest(selector: string, context?: Element): JQuery;
-    /**
-     * For each element in the set, get the first element that matches the selector by testing the element itself and traversing up through its ancestors in the DOM tree.
-     * 
-     * @param obj A jQuery object to match elements against.
-     */
-    closest(obj: JQuery): JQuery;
-    /**
-     * For each element in the set, get the first element that matches the selector by testing the element itself and traversing up through its ancestors in the DOM tree.
-     * 
-     * @param element An element to match elements against.
-     */
-    closest(element: Element): JQuery;
-
-    /**
-     * Get an array of all the elements and selectors matched against the current element up through the DOM tree.
-     * 
-     * @param selectors An array or string containing a selector expression to match elements against (can also be a jQuery object).
-     * @param context A DOM element within which a matching element may be found. If no context is passed in then the context of the jQuery set will be used instead.
-     */
-    closest(selectors: any, context?: Element): any[];
-
-    /**
-     * Get the children of each element in the set of matched elements, including text and comment nodes.
-     */
-    contents(): JQuery;
-
-    /**
-     * End the most recent filtering operation in the current chain and return the set of matched elements to its previous state.
-     */
-    end(): JQuery;
-
-    /**
-     * Reduce the set of matched elements to the one at the specified index.
-     * 
-     * @param index An integer indicating the 0-based position of the element. OR An integer indicating the position of the element, counting backwards from the last element in the set.
-     *  
-     */
-    eq(index: number): JQuery;
-
-    /**
-     * Reduce the set of matched elements to those that match the selector or pass the function's test.
-     * 
-     * @param selector A string containing a selector expression to match the current set of elements against.
-     */
-    filter(selector: string): JQuery;
-    /**
-     * Reduce the set of matched elements to those that match the selector or pass the function's test.
-     * 
-     * @param func A function used as a test for each element in the set. this is the current DOM element.
-     */
-    filter(func: (index: number, element: Element) => any): JQuery;
-    /**
-     * Reduce the set of matched elements to those that match the selector or pass the function's test.
-     * 
-     * @param element An element to match the current set of elements against.
-     */
-    filter(element: Element): JQuery;
-    /**
-     * Reduce the set of matched elements to those that match the selector or pass the function's test.
-     * 
-     * @param obj An existing jQuery object to match the current set of elements against.
-     */
-    filter(obj: JQuery): JQuery;
-
-    /**
-     * Get the descendants of each element in the current set of matched elements, filtered by a selector, jQuery object, or element.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    find(selector: string): JQuery;
-    /**
-     * Get the descendants of each element in the current set of matched elements, filtered by a selector, jQuery object, or element.
-     * 
-     * @param element An element to match elements against.
-     */
-    find(element: Element): JQuery;
-    /**
-     * Get the descendants of each element in the current set of matched elements, filtered by a selector, jQuery object, or element.
-     * 
-     * @param obj A jQuery object to match elements against.
-     */
-    find(obj: JQuery): JQuery;
-
-    /**
-     * Reduce the set of matched elements to the first in the set.
-     */
-    first(): JQuery;
-
-    /**
-     * Reduce the set of matched elements to those that have a descendant that matches the selector or DOM element.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    has(selector: string): JQuery;
-    /**
-     * Reduce the set of matched elements to those that have a descendant that matches the selector or DOM element.
-     * 
-     * @param contained A DOM element to match elements against.
-     */
-    has(contained: Element): JQuery;
-
-    /**
-     * Check the current matched set of elements against a selector, element, or jQuery object and return true if at least one of these elements matches the given arguments.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    is(selector: string): boolean;
-    /**
-     * Check the current matched set of elements against a selector, element, or jQuery object and return true if at least one of these elements matches the given arguments.
-     * 
-     * @param func A function used as a test for the set of elements. It accepts one argument, index, which is the element's index in the jQuery collection.Within the function, this refers to the current DOM element.
-     */
-    is(func: (index: number, element: Element) => boolean): boolean;
-    /**
-     * Check the current matched set of elements against a selector, element, or jQuery object and return true if at least one of these elements matches the given arguments.
-     * 
-     * @param obj An existing jQuery object to match the current set of elements against.
-     */
-    is(obj: JQuery): boolean;
-    /**
-     * Check the current matched set of elements against a selector, element, or jQuery object and return true if at least one of these elements matches the given arguments.
-     * 
-     * @param elements One or more elements to match the current set of elements against.
-     */
-    is(elements: any): boolean;
-
-    /**
-     * Reduce the set of matched elements to the final one in the set.
-     */
-    last(): JQuery;
-
-    /**
-     * Pass each element in the current matched set through a function, producing a new jQuery object containing the return values.
-     * 
-     * @param callback A function object that will be invoked for each element in the current set.
-     */
-    map(callback: (index: number, domElement: Element) => any): JQuery;
-
-    /**
-     * Get the immediately following sibling of each element in the set of matched elements. If a selector is provided, it retrieves the next sibling only if it matches that selector.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    next(selector?: string): JQuery;
-
-    /**
-     * Get all following siblings of each element in the set of matched elements, optionally filtered by a selector.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    nextAll(selector?: string): JQuery;
-
-    /**
-     * Get all following siblings of each element up to but not including the element matched by the selector, DOM node, or jQuery object passed.
-     * 
-     * @param selector A string containing a selector expression to indicate where to stop matching following sibling elements.
-     * @param filter A string containing a selector expression to match elements against.
-     */
-    nextUntil(selector?: string, filter?: string): JQuery;
-    /**
-     * Get all following siblings of each element up to but not including the element matched by the selector, DOM node, or jQuery object passed.
-     * 
-     * @param element A DOM node or jQuery object indicating where to stop matching following sibling elements.
-     * @param filter A string containing a selector expression to match elements against.
-     */
-    nextUntil(element?: Element, filter?: string): JQuery;
-    /**
-     * Get all following siblings of each element up to but not including the element matched by the selector, DOM node, or jQuery object passed.
-     * 
-     * @param obj A DOM node or jQuery object indicating where to stop matching following sibling elements.
-     * @param filter A string containing a selector expression to match elements against.
-     */
-    nextUntil(obj?: JQuery, filter?: string): JQuery;
-
-    /**
-     * Remove elements from the set of matched elements.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    not(selector: string): JQuery;
-    /**
-     * Remove elements from the set of matched elements.
-     * 
-     * @param func A function used as a test for each element in the set. this is the current DOM element.
-     */
-    not(func: (index: number, element: Element) => boolean): JQuery;
-    /**
-     * Remove elements from the set of matched elements.
-     * 
-     * @param elements One or more DOM elements to remove from the matched set.
-     */
-    not(elements: Element|Element[]): JQuery;
-    /**
-     * Remove elements from the set of matched elements.
-     * 
-     * @param obj An existing jQuery object to match the current set of elements against.
-     */
-    not(obj: JQuery): JQuery;
-
-    /**
-     * Get the closest ancestor element that is positioned.
-     */
-    offsetParent(): JQuery;
-
-    /**
-     * Get the parent of each element in the current set of matched elements, optionally filtered by a selector.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    parent(selector?: string): JQuery;
-
-    /**
-     * Get the ancestors of each element in the current set of matched elements, optionally filtered by a selector.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    parents(selector?: string): JQuery;
-
-    /**
-     * Get the ancestors of each element in the current set of matched elements, up to but not including the element matched by the selector, DOM node, or jQuery object.
-     * 
-     * @param selector A string containing a selector expression to indicate where to stop matching ancestor elements.
-     * @param filter A string containing a selector expression to match elements against.
-     */
-    parentsUntil(selector?: string, filter?: string): JQuery;
-    /**
-     * Get the ancestors of each element in the current set of matched elements, up to but not including the element matched by the selector, DOM node, or jQuery object.
-     * 
-     * @param element A DOM node or jQuery object indicating where to stop matching ancestor elements.
-     * @param filter A string containing a selector expression to match elements against.
-     */
-    parentsUntil(element?: Element, filter?: string): JQuery;
-    /**
-     * Get the ancestors of each element in the current set of matched elements, up to but not including the element matched by the selector, DOM node, or jQuery object.
-     * 
-     * @param obj A DOM node or jQuery object indicating where to stop matching ancestor elements.
-     * @param filter A string containing a selector expression to match elements against.
-     */
-    parentsUntil(obj?: JQuery, filter?: string): JQuery;
-
-    /**
-     * Get the immediately preceding sibling of each element in the set of matched elements, optionally filtered by a selector.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    prev(selector?: string): JQuery;
-
-    /**
-     * Get all preceding siblings of each element in the set of matched elements, optionally filtered by a selector.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    prevAll(selector?: string): JQuery;
-
-    /**
-     * Get all preceding siblings of each element up to but not including the element matched by the selector, DOM node, or jQuery object.
-     * 
-     * @param selector A string containing a selector expression to indicate where to stop matching preceding sibling elements.
-     * @param filter A string containing a selector expression to match elements against.
-     */
-    prevUntil(selector?: string, filter?: string): JQuery;
-    /**
-     * Get all preceding siblings of each element up to but not including the element matched by the selector, DOM node, or jQuery object.
-     * 
-     * @param element A DOM node or jQuery object indicating where to stop matching preceding sibling elements.
-     * @param filter A string containing a selector expression to match elements against.
-     */
-    prevUntil(element?: Element, filter?: string): JQuery;
-    /**
-     * Get all preceding siblings of each element up to but not including the element matched by the selector, DOM node, or jQuery object.
-     * 
-     * @param obj A DOM node or jQuery object indicating where to stop matching preceding sibling elements.
-     * @param filter A string containing a selector expression to match elements against.
-     */
-    prevUntil(obj?: JQuery, filter?: string): JQuery;
-
-    /**
-     * Get the siblings of each element in the set of matched elements, optionally filtered by a selector.
-     * 
-     * @param selector A string containing a selector expression to match elements against.
-     */
-    siblings(selector?: string): JQuery;
-
-    /**
-     * Reduce the set of matched elements to a subset specified by a range of indices.
-     * 
-     * @param start An integer indicating the 0-based position at which the elements begin to be selected. If negative, it indicates an offset from the end of the set.
-     * @param end An integer indicating the 0-based position at which the elements stop being selected. If negative, it indicates an offset from the end of the set. If omitted, the range continues until the end of the set.
-     */
-    slice(start: number, end?: number): JQuery;
-
-    /**
-     * Show the queue of functions to be executed on the matched elements.
-     * 
-     * @param queueName A string containing the name of the queue. Defaults to fx, the standard effects queue.
-     */
-    queue(queueName?: string): any[];
-    /**
-     * Manipulate the queue of functions to be executed, once for each matched element.
-     * 
-     * @param newQueue An array of functions to replace the current queue contents.
-     */
-    queue(newQueue: Function[]): JQuery;
-    /**
-     * Manipulate the queue of functions to be executed, once for each matched element.
-     * 
-     * @param callback The new function to add to the queue, with a function to call that will dequeue the next item.
-     */
-    queue(callback: Function): JQuery;
-    /**
-     * Manipulate the queue of functions to be executed, once for each matched element.
-     * 
-     * @param queueName A string containing the name of the queue. Defaults to fx, the standard effects queue.
-     * @param newQueue An array of functions to replace the current queue contents.
-     */
-    queue(queueName: string, newQueue: Function[]): JQuery;
-    /**
-     * Manipulate the queue of functions to be executed, once for each matched element.
-     * 
-     * @param queueName A string containing the name of the queue. Defaults to fx, the standard effects queue.
-     * @param callback The new function to add to the queue, with a function to call that will dequeue the next item.
-     */
-    queue(queueName: string, callback: Function): JQuery;
-}
-declare module "jquery" {
-    export = $;
-}
-declare var jQuery: JQueryStatic;
-declare var $: JQueryStatic;
